Lists all modules available to be loaded

Lists all modules available to be loaded

Usage

module_avail(filter = "")

Arguments

filter

If provided, a character scalar will be used to filter the results. Only modules containing filter as a substring will be returned.

Value

A character vector whose entries correspond to available to be loaded

Examples

module_avail()
#>   [1] "7zip/15.14.1"                           
#>   [2] "AlphaLink/1.0"                          
#>   [3] "AreTomo/1.3.4"                          
#>   [4] "CUDA8/caffe2/0.8.1"                     
#>   [5] "CUDA8/tensorflow/1.1.0"                 
#>   [6] "CUDA8/tensorflow/1.3.0rc0"              
#>   [7] "CUnit/2.1-3"                            
#>   [8] "EMAN/2.3"                               
#>   [9] "EMAN2/2.9"                              
#>  [10] "EMBuilder/1.0"                          
#>  [11] "EPU_group_AFIS/0.3"                     
#>  [12] "FastFold/0.2.0"                         
#>  [13] "Gctf/1.06"                              
#>  [14] "Gctf/1.18cu10.1"                        
#>  [15] "Gctf/1.18cu7.5"                         
#>  [16] "Gctf/1.18cu8.0"                         
#>  [17] "Gctf/1.18cu9.2"                         
#>  [18] "H5toXDS/1.1.0"                          
#>  [19] "HiC-Pro/2.11.4"                         
#>  [20] "IGV/2.10.3"                             
#>  [21] "IGV/2.17.4"                             
#>  [22] "IGV/2.5.2"                              
#>  [23] "IMOD/4.11.5"                            
#>  [24] "IRFinder/1.2.0"                         
#>  [25] "ImageMagick/6.9.11-22"                  
#>  [26] "ImageMagick/7.0.9-5"                    
#>  [27] "JAGS/4.0.0"                             
#>  [28] "JAGS/4.3.0"                             
#>  [29] "ModelAngelo/0.0.1"                      
#>  [30] "MultiQC/1.10.1"                         
#>  [31] "MultiQC/gmon.out"                       
#>  [32] "NAMD-mpich/2.12"                        
#>  [33] "NAMD-mpich/2.14"                        
#>  [34] "NAMD-multicore-gpu/2.14"                
#>  [35] "NAMD-multicore/2.12"                    
#>  [36] "NAMD-multicore/2.14"                    
#>  [37] "NAMD-openmpi/2.12"                      
#>  [38] "NAMD/2.12"                              
#>  [39] "NAMD/2.13nb"                            
#>  [40] "OpenFold/2.0.0"                         
#>  [41] "R/3.3.3"                                
#>  [42] "R/3.4.4"                                
#>  [43] "R/3.5.2"                                
#>  [44] "R/3.5.3"                                
#>  [45] "R/3.6.0"                                
#>  [46] "R/3.6.0m"                               
#>  [47] "R/3.6.1"                                
#>  [48] "R/3.6.1m"                               
#>  [49] "R/3.6.2"                                
#>  [50] "R/3.6.3"                                
#>  [51] "R/4.0.0"                                
#>  [52] "R/4.0.1"                                
#>  [53] "R/4.0.2"                                
#>  [54] "R/4.0.3"                                
#>  [55] "R/4.0.4"                                
#>  [56] "R/4.0.5"                                
#>  [57] "R/4.1.0"                                
#>  [58] "R/4.1.1"                                
#>  [59] "R/4.1.2"                                
#>  [60] "R/4.1.3"                                
#>  [61] "R/4.2.0"                                
#>  [62] "R/4.2.1"                                
#>  [63] "R/4.2.2"                                
#>  [64] "R/4.2.3"                                
#>  [65] "R/4.3.0"                                
#>  [66] "R/4.3.1"                                
#>  [67] "R/4.3.2"                                
#>  [68] "R/4.3.3"                                
#>  [69] "R/4.4.0"                                
#>  [70] "R/MKL/4.3.1"                            
#>  [71] "R/MKL/4.4.0"                            
#>  [72] "R/openBLAS/3.6.2"                       
#>  [73] "R/openBLAS/4.2.1"                       
#>  [74] "R/openBLAS/4.2.2"                       
#>  [75] "R/openBLAS/4.2.3"                       
#>  [76] "R/openBLAS/4.3.0"                       
#>  [77] "R/openBLAS/4.3.1"                       
#>  [78] "R/openBLAS/4.3.2"                       
#>  [79] "R/openBLAS/4.3.3"                       
#>  [80] "R/openBLAS/4.4.0"                       
#>  [81] "RepeatMaker/4.1.2"                      
#>  [82] "SNPsplit/0.2.0"                         
#>  [83] "SPHIRE/1.1"                             
#>  [84] "SPHIRE/1.3"                             
#>  [85] "STAR/2.6.1c"                            
#>  [86] "STAR/2.7.0a"                            
#>  [87] "STAR/2.7.1a"                            
#>  [88] "STAR/2.7.3a"                            
#>  [89] "STAR/2.7.9a"                            
#>  [90] "TRF/4.09.1"                             
#>  [91] "Xvfb/1.19.5"                            
#>  [92] "adxv/1.9.12"                            
#>  [93] "alphafold/2.0.0"                        
#>  [94] "alphafold/2.0.1"                        
#>  [95] "alphafold/2.0.2"                        
#>  [96] "alphafold/2.1.1.0"                      
#>  [97] "alphafold/2.1.2.0"                      
#>  [98] "alphafold/2.2.0.0"                      
#>  [99] "alphafold/2.3.1"                        
#> [100] "alphafold/2.3.2"                        
#> [101] "anaconda3/2019.03"                      
#> [102] "anaconda3/2020.07"                      
#> [103] "anaconda3/4.0.0"                        
#> [104] "anaconda3/4.3.1"                        
#> [105] "anaconda3/latest"                       
#> [106] "annovar/2015-12-14"                     
#> [107] "apache-ant/1.9.7"                       
#> [108] "apptainer/1.0.0"                        
#> [109] "apptainer/1.1.0"                        
#> [110] "apptainer/1.1.8"                        
#> [111] "apptainer/1.1.9"                        
#> [112] "apptainer/1.2.3"                        
#> [113] "apptainer/1.2.3s"                       
#> [114] "aspera/3.5.4"                           
#> [115] "aspera/3.9.1"                           
#> [116] "aspera/3.9.6"                           
#> [117] "autoPROC/20211020"                      
#> [118] "autoPROC/20220608"                      
#> [119] "autoPROC/20230222"                      
#> [120] "autoPROC/20230726"                      
#> [121] "autoSHARP/2.8"                          
#> [122] "awscli/1.22.89"                         
#> [123] "awscli/2.1.25"                          
#> [124] "awscli/2.5.2"                           
#> [125] "axel/2.17.10"                           
#> [126] "bamUtil/1.0.14"                         
#> [127] "bamtools/2.4.1"                         
#> [128] "bazel/0.26.1"                           
#> [129] "bazel/1.2.1"                            
#> [130] "bcftools/1.12"                          
#> [131] "bcftools/1.13"                          
#> [132] "bcftools/1.14"                          
#> [133] "bcftools/1.15"                          
#> [134] "bcftools/1.16"                          
#> [135] "bcftools/1.17"                          
#> [136] "bcftools/1.19"                          
#> [137] "bcftools/1.20"                          
#> [138] "bcftools/1.3.1"                         
#> [139] "bcftools/1.6"                           
#> [140] "bcftools/1.7"                           
#> [141] "bcftools/1.9"                           
#> [142] "bcl-convert/3.10.5"                     
#> [143] "bcl-convert/3.9.3"                      
#> [144] "bcl2fastq/2.19.1"                       
#> [145] "bcl2fastq/2.20.0"                       
#> [146] "beast/1.8.3"                            
#> [147] "beast2/2.4.0"                           
#> [148] "bedops/2.4.26"                          
#> [149] "bedtools/2.25.0"                        
#> [150] "bedtools/2.26.0"                        
#> [151] "bedtools/2.31.1"                        
#> [152] "binutils/2.35.2-gcc-4.8.5"              
#> [153] "binutils/2.35.2-gcc-9.1.0"              
#> [154] "biobambam2/2.0.182-gcc9.1.0"            
#> [155] "bismark/0.16.1"                         
#> [156] "bismark/0.19.0"                         
#> [157] "bismark/0.19.1"                         
#> [158] "bismark/0.20.0"                         
#> [159] "boost/1.55.0"                           
#> [160] "boost/1.58.0"                           
#> [161] "boost/1.60.0"                           
#> [162] "boost/1.66.0"                           
#> [163] "boost/1.72.0-gcc-11.3.0"                
#> [164] "bowtie/1.1.2"                           
#> [165] "bowtie2/2.2.9"                          
#> [166] "bowtie2/2.3.2"                          
#> [167] "bowtie2/2.3.4.1"                        
#> [168] "bowtie2/2.4.4"                          
#> [169] "bpipe/0.9.9.2"                          
#> [170] "breakdancer/1.1.2"                      
#> [171] "breakdancer/1.4.5"                      
#> [172] "bsoft/2.1.0"                            
#> [173] "buster/20211020"                        
#> [174] "buster/20220608"                        
#> [175] "buster/20230217"                        
#> [176] "bwa/0.7.13"                             
#> [177] "bwa/0.7.15"                             
#> [178] "bwa/0.7.17"                             
#> [179] "bwtool/1.0"                             
#> [180] "bzip2/1.0.6"                            
#> [181] "caffe2/0.8.1"                           
#> [182] "canu/1.3"                               
#> [183] "canu/1.5"                               
#> [184] "cbflib/0.9.6"                           
#> [185] "ccp4/7.1"                               
#> [186] "ccp4/8.0"                               
#> [187] "ccpem/1.5.0"                            
#> [188] "cellprofiler/4.2.1"                     
#> [189] "cellprofiler/4.2.4"                     
#> [190] "cellranger-arc/1.0.1"                   
#> [191] "cellranger-arc/2.0.2"                   
#> [192] "cellranger-atac/1.0.0"                  
#> [193] "cellranger-atac/1.2.0"                  
#> [194] "cellranger-atac/2.1.0"                  
#> [195] "cellranger/2.2.0"                       
#> [196] "cellranger/3.0.1"                       
#> [197] "cellranger/3.0.2"                       
#> [198] "cellranger/3.1.0"                       
#> [199] "cellranger/4.0.0"                       
#> [200] "cellranger/5.0.0"                       
#> [201] "cellranger/6.0.0"                       
#> [202] "cellranger/6.1.2"                       
#> [203] "cellranger/7.0.0"                       
#> [204] "cellranger/7.1.0"                       
#> [205] "cellranger/8.0.0"                       
#> [206] "cellsnp-lite/1.2.0"                     
#> [207] "chimera/1.11"                           
#> [208] "chimerax/0.8"                           
#> [209] "chimerax/0.9"                           
#> [210] "chimerax/0.93"                          
#> [211] "chimerax/1.1"                           
#> [212] "chimerax/1.3"                           
#> [213] "circos/0.69"                            
#> [214] "circos/0.69-9"                          
#> [215] "cistem/1.0.0-beta"                      
#> [216] "clara-parabricks/4.0.1-1"               
#> [217] "clara-parabricks/4.3.0-1"               
#> [218] "clover/2015-08-28"                      
#> [219] "cluster-utils/18.08.1"                  
#> [220] "cmake/3.25.1"                           
#> [221] "cnvnator/0.3.3"                         
#> [222] "computecpp/1.0.5"                       
#> [223] "coot/0.8.8"                             
#> [224] "coot/0.8.9"                             
#> [225] "coot/0.8.9.1"                           
#> [226] "crYOLO/1.9.2"                           
#> [227] "crYOLO/1.9.7"                           
#> [228] "cryoEF/1.1.0"                           
#> [229] "cryoem-utils/0.1.0"                     
#> [230] "ctffind/4.1.13"                         
#> [231] "ctffind/4.1.14"                         
#> [232] "ctffind/4.1.5"                          
#> [233] "cuda/10.0"                              
#> [234] "cuda/10.1"                              
#> [235] "cuda/10.2"                              
#> [236] "cuda/11.0"                              
#> [237] "cuda/11.1"                              
#> [238] "cuda/11.2"                              
#> [239] "cuda/11.3.1"                            
#> [240] "cuda/11.4"                              
#> [241] "cuda/11.5.1"                            
#> [242] "cuda/11.6.2"                            
#> [243] "cuda/11.7.1"                            
#> [244] "cuda/11.8"                              
#> [245] "cuda/12.2"                              
#> [246] "cuda/8.0"                               
#> [247] "cuda/9.2"                               
#> [248] "cudnn/10.1-7.6.5"                       
#> [249] "cudnn/11.2-8.1.1.33"                    
#> [250] "cudnn/11.3.1-8.4.0"                     
#> [251] "cudnn/11.3.1-8.6.0.162"                 
#> [252] "cudnn/11.3.1-8.9.5"                     
#> [253] "cudnn/11.x-8.9.7.29"                    
#> [254] "cudnn/11.x-9.0.0.312"                   
#> [255] "cudnn/12.x-8.9.7.29"                    
#> [256] "curl/7.65.0"                            
#> [257] "cutadapt/2.9"                           
#> [258] "cutadapt/3.4"                           
#> [259] "cutadapt/3.7"                           
#> [260] "cutadapt/4.0"                           
#> [261] "dalilite/5.1"                           
#> [262] "darknet/61c9d02cu10.0"                  
#> [263] "dectris-neggia/1.2.0"                   
#> [264] "deepEMhancer/0.15"                      
#> [265] "deeptools/3.5.1"                        
#> [266] "demuxlet/noversion"                     
#> [267] "depot_tools/6c7b829"                    
#> [268] "dials/3.10.0"                           
#> [269] "dials/3.12.0"                           
#> [270] "dials/3.17.0"                           
#> [271] "dorado/0.1.1-gcc-9.1.0"                 
#> [272] "dorado/0.2.1-gcc-9.1.0"                 
#> [273] "dorado/0.3.0"                           
#> [274] "dorado/0.3.2"                           
#> [275] "dorado/0.3.4"                           
#> [276] "dorado/0.4.1"                           
#> [277] "dorado/0.5.0"                           
#> [278] "dorado/0.5.2"                           
#> [279] "dorado/6.0.0"                           
#> [280] "dotnet/2.1.809"                         
#> [281] "dotnet/3.1.412"                         
#> [282] "dotnet/6.0.408"                         
#> [283] "dotnet/7.0.305"                         
#> [284] "doublecmd/0.9.10.gtk2"                  
#> [285] "dua-cli/2.17.8"                         
#> [286] "dua-cli/2.19.0"                         
#> [287] "duck/6.9.3"                             
#> [288] "duckdb/0.6.1"                           
#> [289] "edirect/15.0"                           
#> [290] "egaclient/2.2.2"                        
#> [291] "eigen/3.4.0"                            
#> [292] "eiger2cbf/160415"                       
#> [293] "eiger2cbf/latest"                       
#> [294] "elbencho/1.7-1cu10.1"                   
#> [295] "ensembl-vep/107.0"                      
#> [296] "ensembl-vep/89.0"                       
#> [297] "ensembl-vep/89.8"                       
#> [298] "ensembl-vep/98.3"                       
#> [299] "evince/3.28.2"                          
#> [300] "fastqc/0.11.5"                          
#> [301] "fastqc/0.11.8"                          
#> [302] "fastx-toolkit/0.0.13"                   
#> [303] "feh/3.6.3"                              
#> [304] "ffmpeg/3.4"                             
#> [305] "fftw/3.3.9"                             
#> [306] "fio/3.16"                               
#> [307] "fio/3.19"                               
#> [308] "flash/1.2.11"                           
#> [309] "fluxbox/1.3.7"                          
#> [310] "freebayes/1.1.0"                        
#> [311] "freebayes/1.2.0"                        
#> [312] "gatk/4.2.4.1"                           
#> [313] "gatk/4.2.5.0"                           
#> [314] "gautomatch/0.56cu10.1"                  
#> [315] "gautomatch/0.56cu8.0"                   
#> [316] "gcc/10.3.0"                             
#> [317] "gcc/11.1.0"                             
#> [318] "gcc/11.2.0"                             
#> [319] "gcc/11.3.0"                             
#> [320] "gcc/12.2.0"                             
#> [321] "gcc/6.3.0"                              
#> [322] "gcc/7.1.0"                              
#> [323] "gcc/7.3.0"                              
#> [324] "gcc/8.1.0"                              
#> [325] "gcc/8.2.0"                              
#> [326] "gcc/8.3.0"                              
#> [327] "gcc/9.1.0"                              
#> [328] "gdal/2.4.4"                             
#> [329] "gdal/3.6.2"                             
#> [330] "gem/3.0"                                
#> [331] "gemini/0.20.1"                          
#> [332] "gencore/0.17.2"                         
#> [333] "genetorrent/3.8.7"                      
#> [334] "genozip/12.0.11"                        
#> [335] "geos/3.7.0"                             
#> [336] "germline/1.5.1"                         
#> [337] "gh-cli/2.0"                             
#> [338] "gh-cli/2.49.2"                          
#> [339] "git/2.37.1"                             
#> [340] "glew/1.9.0"                             
#> [341] "glib/1.2.10"                            
#> [342] "glpk/4.65"                              
#> [343] "glpk/5.0"                               
#> [344] "gmap-gsnap/2017-05-08"                  
#> [345] "gmap-gsnap/2017-06-16"                  
#> [346] "gmp/6.3.0"                              
#> [347] "gnuplot/5.2.6"                          
#> [348] "go/1.13.15"                             
#> [349] "go/1.19.4"                              
#> [350] "go/1.20.2"                              
#> [351] "google-cloud-sdk/194.0.0"               
#> [352] "google-cloud-sdk/348.0.0"               
#> [353] "google-cloud-sdk/437.0.1"               
#> [354] "gradle/4.3"                             
#> [355] "graphmap/0.3.1"                         
#> [356] "graphmap/0.5.1"                         
#> [357] "graphmap/0.5.2"                         
#> [358] "gridss/2.11.1"                          
#> [359] "gridss/2.13.2"                          
#> [360] "gromacs/2021.2"                         
#> [361] "gromacs/2023.3"                         
#> [362] "gromacs/2023.3_cuda11.7.1"              
#> [363] "gromacs/4.5.5_cryo_fit"                 
#> [364] "groovy/4.0.0"                           
#> [365] "gsl/2.4"                                
#> [366] "gsl/2.6"                                
#> [367] "guppy-cpu/2.3.7"                        
#> [368] "guppy-cpu/3.1.5"                        
#> [369] "guppy-cpu/3.2.1"                        
#> [370] "guppy-cpu/3.2.2"                        
#> [371] "guppy-cpu/3.2.4"                        
#> [372] "guppy-cpu/3.3.0"                        
#> [373] "guppy-cpu/3.6.1"                        
#> [374] "guppy-cpu/4.0.15"                       
#> [375] "guppy-gpu/2.3.1"                        
#> [376] "guppy-gpu/2.3.7"                        
#> [377] "guppy-gpu/3.1.5"                        
#> [378] "guppy-gpu/3.2.1"                        
#> [379] "guppy-gpu/3.2.2"                        
#> [380] "guppy-gpu/3.2.4"                        
#> [381] "guppy-gpu/3.3.0"                        
#> [382] "guppy-gpu/3.6.1"                        
#> [383] "guppy-gpu/4.0.15"                       
#> [384] "guppy-gpu/6.2.1"                        
#> [385] "guppy-gpu/6.3.7"                        
#> [386] "gzip/1.10"                              
#> [387] "haplopainter/1.043"                     
#> [388] "hdf4/4.2.11"                            
#> [389] "hdf5-mpich/1.10.5_3.3"                  
#> [390] "hdf5/1.10.5"                            
#> [391] "hdf5/1.12.2"                            
#> [392] "hdf5/1.14.0"                            
#> [393] "hdf5/1.8.16"                            
#> [394] "hdf5/1.8.20"                            
#> [395] "hdf5/1.8.21"                            
#> [396] "hdfview/2.11.0"                         
#> [397] "hisat2/2.0.3"                           
#> [398] "hisat2/2.0.5"                           
#> [399] "hisat2/2.2.1"                           
#> [400] "hmmer/2.3.2"                            
#> [401] "hmmer/3.1b2"                            
#> [402] "hmmer/3.1b2-mpi"                        
#> [403] "hmmer/3.2.1"                            
#> [404] "hmmer/3.3.2"                            
#> [405] "homer/4.11"                             
#> [406] "hpl/2.3"                                
#> [407] "htop/2.0.2"                             
#> [408] "htslib/1.10.2"                          
#> [409] "htslib/1.12"                            
#> [410] "htslib/1.13"                            
#> [411] "htslib/1.14"                            
#> [412] "htslib/1.15"                            
#> [413] "htslib/1.17"                            
#> [414] "htslib/1.19"                            
#> [415] "htslib/1.19.1"                          
#> [416] "htslib/1.20"                            
#> [417] "htslib/1.3.1"                           
#> [418] "htslib/1.3.2"                           
#> [419] "htslib/1.4.1"                           
#> [420] "htslib/1.5"                             
#> [421] "htslib/1.6"                             
#> [422] "htslib/1.7"                             
#> [423] "htslib/1.9"                             
#> [424] "icewm/2.8.0"                            
#> [425] "iftop/1.0"                              
#> [426] "imp/2.5.0"                              
#> [427] "intel-ipp/2019.5.281"                   
#> [428] "intel-mkl/2019.3.199"                   
#> [429] "intel-mpi/2019.3.199"                   
#> [430] "intel-tbb/2019.3.199"                   
#> [431] "intel_mkl_2019/2019.5.075"              
#> [432] "intel_mpi_2019/2019.5.075"              
#> [433] "intel_mpi_2019/default_pipeline.star"   
#> [434] "ior-slurm/3.2.1mpich3.3"                
#> [435] "ior-slurm/3.2.1openMPI4.0.2"            
#> [436] "ior/3.2.1mpio4.0.1"                     
#> [437] "iozone/3.491"                           
#> [438] "java/1.7.0_79"                          
#> [439] "java/1.8.0_131"                         
#> [440] "java/1.8.0_151"                         
#> [441] "java/1.8.0_211"                         
#> [442] "java/1.8.0_92"                          
#> [443] "jbrowse-desktop/2.3.2"                  
#> [444] "jellyfish/2.3.1"                        
#> [445] "joe/4.6"                                
#> [446] "joe/n"                                  
#> [447] "julia/0.6.4"                            
#> [448] "julia/1.0.1"                            
#> [449] "julia/1.5.3"                            
#> [450] "julia/1.7.3"                            
#> [451] "julia/1.9.2"                            
#> [452] "juncBASE/1.2-beta"                      
#> [453] "jupyter-config/1.0"                     
#> [454] "jupyter/latest"                         
#> [455] "jupyterlab/latest"                      
#> [456] "kallisto/0.44.0"                        
#> [457] "kallisto/0.46.0"                        
#> [458] "kraken/2.1.1"                           
#> [459] "kraken/2.1.2"                           
#> [460] "lapack/3.8.0"                           
#> [461] "last/864"                               
#> [462] "libaio/0.3.111"                         
#> [463] "libglvnd/1.1.1"                         
#> [464] "libiconv/1.16"                          
#> [465] "libmaus2/2.0.798-gcc9.1.0"              
#> [466] "libtorch-gpu/1.12.1"                    
#> [467] "lighttpd/1.4.45"                        
#> [468] "lxrandr/0.3.1"                          
#> [469] "lz4/1.9.3"                              
#> [470] "macs2/2.2.7.1"                          
#> [471] "macs3/3.0.0b3"                          
#> [472] "mageck/0.5.9.4"                         
#> [473] "magma-gpu/2.7.0-11.3.1"                 
#> [474] "mariadb-client/10.11.2"                 
#> [475] "mariadb-connector-c/3.1.11"             
#> [476] "mariadb-connector-c/3.3.4"              
#> [477] "maven/3.3.9"                            
#> [478] "maven/3.5.0"                            
#> [479] "maven/3.9.1"                            
#> [480] "mbuffer/20171011"                       
#> [481] "mcl/14-137"                             
#> [482] "mediaflux-data-mover/1.0.11"            
#> [483] "mediaflux-data-mover/1.0.12"            
#> [484] "mediaflux-data-mover/1.0.8"             
#> [485] "mediaflux-data-mover/1.1.15"            
#> [486] "meme/4.12.0"                            
#> [487] "meme/5.0.5"                             
#> [488] "meme/openmpi-4.12.0"                    
#> [489] "meme/openmpi-5.0.5"                     
#> [490] "mercurial/4.8"                          
#> [491] "micromamba/latest"                      
#> [492] "miniconda2/4.5.12"                      
#> [493] "miniconda3-alt/4.10.3"                  
#> [494] "miniconda3/4.10.3"                      
#> [495] "miniconda3/latest"                      
#> [496] "minimap2/2.17"                          
#> [497] "minimap2/2.4"                           
#> [498] "molgenis-compute/17.08.1"               
#> [499] "mono/6.10.0.104"                        
#> [500] "mono/6.12.0.90"                         
#> [501] "motioncor2/1.2.1cu8.0"                  
#> [502] "motioncor2/1.2.1cu9.2"                  
#> [503] "motioncor2/1.2.2cu10.0"                 
#> [504] "motioncor2/1.2.2cu8.0"                  
#> [505] "motioncor2/1.2.2cu9.2"                  
#> [506] "motioncor2/1.2.6cu10.1"                 
#> [507] "motioncor2/1.2.6cu8.0"                  
#> [508] "motioncor2/1.2.6cu9.2"                  
#> [509] "motioncor2/1.3.1cu10.1"                 
#> [510] "motioncor2/1.3.1cu9.2"                  
#> [511] "motioncor2/1.4.0cu10.0"                 
#> [512] "motioncor2/1.4.0cu10.1"                 
#> [513] "motioncor2/1.4.0cu10.2"                 
#> [514] "motioncor2/1.4.0cu11.0"                 
#> [515] "motioncor2/1.4.0cu9.2"                  
#> [516] "motioncor2/1.4.4cu10.0"                 
#> [517] "motioncor2/1.4.4cu10.1"                 
#> [518] "motioncor2/1.4.4cu10.2"                 
#> [519] "motioncor2/1.4.4cu11.1"                 
#> [520] "motioncor2/1.4.4cu11.2"                 
#> [521] "motioncor2/1.4.4cu11.3.1"               
#> [522] "motioncor2/1.5.0cu10.1"                 
#> [523] "motioncor2/1.5.0cu10.2"                 
#> [524] "motioncor2/1.5.0cu11.1"                 
#> [525] "motioncor2/1.5.0cu11.2"                 
#> [526] "motioncor2/1.5.0cu11.4"                 
#> [527] "motioncor2/11.5.0cu11.3.1"              
#> [528] "motioncorr/2.1"                         
#> [529] "mpfr/4.1.0"                             
#> [530] "mpich-slurm/3.3.2"                      
#> [531] "mpich-slurm/3.4.1"                      
#> [532] "mpich-slurm/3.4.2"                      
#> [533] "mpich/3.3"                              
#> [534] "mpich/3.3.2"                            
#> [535] "mrcz/0.1.3"                             
#> [536] "multiSNV/2.3-10-g06102a5"               
#> [537] "muscle/3.8.31"                          
#> [538] "mutect/1.1.7"                           
#> [539] "nanonet-openBLAS/2.0.0"                 
#> [540] "nanonet/2.0.0"                          
#> [541] "nanopolish/0.5.0"                       
#> [542] "nanopolish/0.6.2"                       
#> [543] "nanopolish/0.7.1"                       
#> [544] "nanopolish/0.8.1"                       
#> [545] "nanopolish/0.8.5"                       
#> [546] "nanopore/1.0"                           
#> [547] "ncbi-blast/2.10.1+"                     
#> [548] "ncbi-blast/2.11.0+"                     
#> [549] "ncbi-blast/2.6.0+"                      
#> [550] "ncbi-blast/2.7.1+"                      
#> [551] "ncbi-blast/2.9.0"                       
#> [552] "ncbi-blast/2.9.0a"                      
#> [553] "ncftp/3.2.6"                            
#> [554] "nedit/5.5"                              
#> [555] "netcdf/4.4.1.1"                         
#> [556] "netcdf_c/4.9.2"                         
#> [557] "nextflow-tw-agent/0.5.0"                
#> [558] "nextflow-tw-cli/0.8.0"                  
#> [559] "nextflow/22.04.5"                       
#> [560] "nextflow/22.10.4"                       
#> [561] "nextflow/23.04.2"                       
#> [562] "nextflow/23.10.0"                       
#> [563] "nf-core/2.7.2"                          
#> [564] "ninja/1.10.0"                           
#> [565] "nix/fix"                                
#> [566] "nix/latest"                             
#> [567] "nmap-ncat/7.91"                         
#> [568] "node/6.11.0"                            
#> [569] "node/8.11.2"                            
#> [570] "node/9.9.0"                             
#> [571] "nodejs/10.24.1"                         
#> [572] "nodejs/16.19.0"                         
#> [573] "nodejs/17.9.1"                          
#> [574] "novoalign/3.04.04"                      
#> [575] "novocraft/3.08.02"                      
#> [576] "oases/0.2.09"                           
#> [577] "ocl-icd/2.3.1"                          
#> [578] "octave/6.4.0-gcc11.1.0"                 
#> [579] "oneMKL/2022.1.0.223"                    
#> [580] "openBLAS/0.3.21-gcc-11.1.0"             
#> [581] "openBLAS/0.3.21-gcc-11.1.0-skylake"     
#> [582] "openBLAS/0.3.23-gcc-11.3.0"             
#> [583] "openBLAS/0.3.6-gcc-9.1.0"               
#> [584] "openCV/2.4.13.6"                        
#> [585] "openCV/4.2.0"                           
#> [586] "openMPI-slurm/4.1.0"                    
#> [587] "openMPI-slurm/4.1.2"                    
#> [588] "openMPI-slurm/4.1.5"                    
#> [589] "openMPI-slurm/4.1.5-gcc-11.2.0"         
#> [590] "openMPI-slurm/4.1.6"                    
#> [591] "openMPI-slurm/4.1.6rc2"                 
#> [592] "openMPI/4.1.4"                          
#> [593] "openSSL/1.0.2r"                         
#> [594] "openSSL/1.1.1b"                         
#> [595] "openSSL/1.1.1g"                         
#> [596] "openSSL/1.1.1k"                         
#> [597] "openSSL/1.1.1n"                         
#> [598] "openSSL/1.1.1u"                         
#> [599] "openSSL/3.0.13"                         
#> [600] "openjdk/1.8.0"                          
#> [601] "openjdk/13.0.2"                         
#> [602] "openjdk/14.0.2"                         
#> [603] "openjdk/15.0.2"                         
#> [604] "openjdk/16.0.1"                         
#> [605] "openjdk/17.0.2"                         
#> [606] "openjdk/18.0.2"                         
#> [607] "openjdk/20.0.1"                         
#> [608] "openmpi-CUDA8/3.0.0"                    
#> [609] "openmpi/2.1.0"                          
#> [610] "openmpi/2.1.1"                          
#> [611] "openmpi/3.0.0"                          
#> [612] "openmpi/4.1.1"                          
#> [613] "openmpi/4.1.1-slurm"                    
#> [614] "owncloud-client/2.3.3"                  
#> [615] "pandoc/1.17.0.3"                        
#> [616] "pandoc/2.0.0.1"                         
#> [617] "pandoc/2.14.2"                          
#> [618] "pandoc/2.19.2"                          
#> [619] "pandoc/2.3.1"                           
#> [620] "parallel/20190722"                      
#> [621] "pbs-drmaa/1.0.19"                       
#> [622] "pbstop/1.0"                             
#> [623] "pcre2/10.34-gcc-9.1.0"                  
#> [624] "pcre2/10.36-gcc-11.1.0"                 
#> [625] "pcre2/10.42-gcc-11.3.0"                 
#> [626] "peakseq/1.31"                           
#> [627] "pear/0.9.10"                            
#> [628] "pear/0.9.11"                            
#> [629] "perl/5.22.2"                            
#> [630] "perl/5.26.3"                            
#> [631] "perl/5.30.1"                            
#> [632] "pgsql/15.1"                             
#> [633] "phenix-nightly/1.15-3459"               
#> [634] "phenix-rosetta/1.11.1-2575"             
#> [635] "phenix/1.15-3459"                       
#> [636] "phenix/1.17-3644"                       
#> [637] "phenix/1.18.2-3874"                     
#> [638] "phenix/1.20.1-4487"                     
#> [639] "php/5.6.21"                             
#> [640] "picard-tools/2.26.11"                   
#> [641] "pigz/2.6"                               
#> [642] "pindel/0.2.5b9"                         
#> [643] "pixz/1.0.6"                             
#> [644] "platypus/0.8.1"                         
#> [645] "plink/1.9"                              
#> [646] "pmix/2.2.5"                             
#> [647] "pmix/3.2.3"                             
#> [648] "pmix/4.2.3"                             
#> [649] "poetry/latest"                          
#> [650] "poretools/0.6.0"                        
#> [651] "proj/4.9.3"                             
#> [652] "proj/6.3.2"                             
#> [653] "proj/9.1.0"                             
#> [654] "pyega3/pyega3-3.4.0/bin/pyega3"         
#> [655] "pyega3/pyega3-3.4.0/pyega3-3.4.0.sif"   
#> [656] "pyem/0.5"                               
#> [657] "pymol/1.7.4.0"                          
#> [658] "pymol/2.5.0"                            
#> [659] "python/2.7.18"                          
#> [660] "python/3.10.12"                         
#> [661] "python/3.11.4"                          
#> [662] "python/3.12.0"                          
#> [663] "python/3.5.1"                           
#> [664] "python/3.5.3"                           
#> [665] "python/3.6.5-intel"                     
#> [666] "python/3.7.0"                           
#> [667] "python/3.7.17"                          
#> [668] "python/3.8.3"                           
#> [669] "python/3.8.8"                           
#> [670] "python/3.9.17"                          
#> [671] "pytorch/2.0.1"                          
#> [672] "qiime2/2019.7"                          
#> [673] "qiime2/2020.8"                          
#> [674] "qpdf/10.0.1"                            
#> [675] "quarto/1.1.189"                         
#> [676] "quarto/1.3.450"                         
#> [677] "rc-tools/1.0"                           
#> [678] "rclone/1.55.0"                          
#> [679] "relion-cpu/3.1b-icc"                    
#> [680] "relion-cpu/3.1beta-icc"                 
#> [681] "relion/3.1.3-cu11.2"                    
#> [682] "relion/3.1.3-gcc11.1.0-cu11.4"          
#> [683] "relion/4.0-beta-1-cu10.1"               
#> [684] "relion/4.0-beta-1-cu11.2"               
#> [685] "relion/4.0-cu11.3.1"                    
#> [686] "relion/5.0-beta-cu11.3.1"               
#> [687] "rmblast/2.11.0"                         
#> [688] "rnammer/1.2"                            
#> [689] "root/6.10.00"                           
#> [690] "rsem/1.3.0"                             
#> [691] "rsem/1.3.1"                             
#> [692] "rstudio_singularity/1.0.0"              
#> [693] "ruby/2.3.1"                             
#> [694] "salmon/0.11.2"                          
#> [695] "salmon/0.14.1"                          
#> [696] "salmon/0.9.1"                           
#> [697] "salmon/1.1.0"                           
#> [698] "salmon/1.10.2"                          
#> [699] "salmon/1.5.2"                           
#> [700] "salmon/1.9.0"                           
#> [701] "salmonTE/0.3"                           
#> [702] "sambamba/0.6.6"                         
#> [703] "samblaster/0.1.23"                      
#> [704] "samtools/0.1.18"                        
#> [705] "samtools/0.1.19"                        
#> [706] "samtools/1.10"                          
#> [707] "samtools/1.12"                          
#> [708] "samtools/1.13"                          
#> [709] "samtools/1.14"                          
#> [710] "samtools/1.15"                          
#> [711] "samtools/1.16.1"                        
#> [712] "samtools/1.17"                          
#> [713] "samtools/1.18"                          
#> [714] "samtools/1.19.2"                        
#> [715] "samtools/1.20"                          
#> [716] "samtools/1.3.1"                         
#> [717] "samtools/1.4.1"                         
#> [718] "samtools/1.5"                           
#> [719] "samtools/1.6"                           
#> [720] "samtools/1.7"                           
#> [721] "samtools/1.9"                           
#> [722] "scala/2.12.2"                           
#> [723] "schrodinger/2023-4"                     
#> [724] "scipion/2.0"                            
#> [725] "scipion/3.0"                            
#> [726] "seqmonk/1.45.1"                         
#> [727] "seqmonk/1.46"                           
#> [728] "seqmonk/1.47.0"                         
#> [729] "seqmonk/1.47.1"                         
#> [730] "seqmonk/1.47.2"                         
#> [731] "seqmonk/1.48.0"                         
#> [732] "seqmonk/1.48.1"                         
#> [733] "sidesplitter/1.2"                       
#> [734] "signalp/4.1"                            
#> [735] "singularity/2.5.1"                      
#> [736] "singularity/3.0.3"                      
#> [737] "singularity/3.1.1"                      
#> [738] "singularity/3.3.0"                      
#> [739] "singularity/3.5.0"                      
#> [740] "singularity/3.5.2"                      
#> [741] "singularity/3.6.2"                      
#> [742] "singularity/3.7.4"                      
#> [743] "singularity/3.8.2"                      
#> [744] "singularity/3.8.4"                      
#> [745] "slurm-contribs/20.11.5"                 
#> [746] "snakemake/7.12.0"                       
#> [747] "snpEff/4.3i"                            
#> [748] "snpEff/4.3t"                            
#> [749] "soapdenovo-trans/1.03"                  
#> [750] "soapdenovo2/2.04"                       
#> [751] "spaceranger/1.2.2"                      
#> [752] "spaceranger/1.3.0"                      
#> [753] "spaceranger/2.0.0"                      
#> [754] "spaceranger/2.1.1"                      
#> [755] "spaceranger/3.0.0"                      
#> [756] "spack/0.21.2"                           
#> [757] "spades/3.11.0"                          
#> [758] "spawn-fcgi/1.6.5"                       
#> [759] "splitcode/0.28.3"                       
#> [760] "spp/1.14"                               
#> [761] "sqlite/3.38.5"                          
#> [762] "sqlite/3.40.0"                          
#> [763] "sqlite/3.40.1"                          
#> [764] "squashfuse/0.1.102"                     
#> [765] "squashfuse/0.5.0"                       
#> [766] "sra-toolkit/2.10.9"                     
#> [767] "sra-toolkit/2.8.2-1"                    
#> [768] "sra-toolkit/2.9.6-1"                    
#> [769] "sra-toolkit/3.0.0"                      
#> [770] "sshpass/1.05"                           
#> [771] "stack/1.5.1"                            
#> [772] "stack/1.7.1"                            
#> [773] "stornext/1.1"                           
#> [774] "stubl/0.0.10"                           
#> [775] "subread/1.5.0-p2"                       
#> [776] "subread/1.5.1"                          
#> [777] "subread/1.5.2"                          
#> [778] "subread/1.6.3"                          
#> [779] "subread/2.0.6"                          
#> [780] "summovie/1.0.2"                         
#> [781] "sxgui_settings/gui_settings_project.txt"
#> [782] "tadbit/0.2"                             
#> [783] "tadbit/gmon.out"                        
#> [784] "tar/1.34"                               
#> [785] "tar/1.35"                               
#> [786] "tcpdump/4.9.2"                          
#> [787] "tensorRT/10.0-6.0.1.5"                  
#> [788] "tensorRT/10.1-6.0.1.5"                  
#> [789] "tensorRT/11.2-7.2.3.4"                  
#> [790] "tensorRT/11.6.2-8.6.0.12"               
#> [791] "tensorRT/11.x-8.6.1.6"                  
#> [792] "tensorRT/12.x-8.6.1.6"                  
#> [793] "tensorflow-gpu/10.0-1.15.2"             
#> [794] "tensorflow-gpu/10.1-2.1.0"              
#> [795] "tensorflow-gpu/11.2-2.11.0"             
#> [796] "tensorflow-gpu/12.2-2.15.0"             
#> [797] "texlive/20171130"                       
#> [798] "texlive/3.14159265"                     
#> [799] "tmhmm/2.0c"                             
#> [800] "topaz/0.2.3"                            
#> [801] "topaz/0.2.4"                            
#> [802] "topaz/0.2.5"                            
#> [803] "topaz/0.2.5a"                           
#> [804] "tophat/2.1.1"                           
#> [805] "transdecoder/4.0.0"                     
#> [806] "transdecoder/5.7.1"                     
#> [807] "trickle/1.07"                           
#> [808] "trimgalore/0.4.2"                       
#> [809] "trimgalore/0.4.4"                       
#> [810] "trimgalore/0.6.7"                       
#> [811] "trimmomatic/0.36"                       
#> [812] "trinityrnaseq/2.15.1"                   
#> [813] "trinityrnaseq/2.8.4"                    
#> [814] "trinotate/3.0.1"                        
#> [815] "trinotate/3.0.2"                        
#> [816] "turbovnc/3.0.1"                         
#> [817] "tvips/0.0.3"                            
#> [818] "ucsc-tools/331"                         
#> [819] "ultraplex/1.2.5"                        
#> [820] "umi-tools/1.1.4"                        
#> [821] "unblur/1.0.2"                           
#> [822] "usearch/7.0.1090"                       
#> [823] "usearch/8.1.1861"                       
#> [824] "usearch/9.2.64"                         
#> [825] "v8/3.14.5.10"                           
#> [826] "vardict/1.5.1"                          
#> [827] "varscan/2.3.9"                          
#> [828] "vast-trino/375"                         
#> [829] "vcflib/1.0.0-rc1"                       
#> [830] "vcftools/0.1.13"                        
#> [831] "velvet/1.2.10"                          
#> [832] "vim/9.0.1679"                           
#> [833] "vireoSNP/0.5.6"                         
#> [834] "vireoSNP/0.5.8"                         
#> [835] "vsearch/2.4.3"                          
#> [836] "websockify/0.11.0"                      
#> [837] "websockify/0.9.0"                       
#> [838] "wine/6.9"                               
#> [839] "wine/7.17"                              
#> [840] "xds/latest"                             
#> [841] "xdsgui/2017-01-25"                      
#> [842] "xdsgui/2020-12-02"                      
#> [843] "xenomapper/1.0.1"                       
#> [844] "xz/5.2.2"                               
#> [845] "zeromq/4.3.4"                           
#> [846] "zlib/1.2.11"                            
#> [847] "zlib/1.2.8"                             
#> [848] "zstd/1.5.0"