Lists all modules available to be loaded

Lists all modules available to be loaded

Usage

module_avail(filter = "")

Arguments

filter

If provided, a character scalar will be used to filter the results. Only modules containing filter as a substring will be returned.

Value

A character vector whose entries correspond to available to be loaded

Examples

module_avail()
#>   [1] "7zip/15.14.1"                           
#>   [2] "CUDA8/caffe2/0.8.1"                     
#>   [3] "CUDA8/tensorflow/1.1.0"                 
#>   [4] "CUDA8/tensorflow/1.3.0rc0"              
#>   [5] "CUnit/2.1-3"                            
#>   [6] "EMAN/2.3"                               
#>   [7] "EMAN2/2.9"                              
#>   [8] "EMBuilder/1.0"                          
#>   [9] "EPU_group_AFIS/0.3"                     
#>  [10] "Gctf/1.06"                              
#>  [11] "Gctf/1.18cu10.1"                        
#>  [12] "Gctf/1.18cu7.5"                         
#>  [13] "Gctf/1.18cu8.0"                         
#>  [14] "Gctf/1.18cu9.2"                         
#>  [15] "H5toXDS/1.1.0"                          
#>  [16] "HiC-Pro/2.11.4"                         
#>  [17] "IGV/2.10.3"                             
#>  [18] "IGV/2.5.2"                              
#>  [19] "IMOD/4.11.5"                            
#>  [20] "IRFinder/1.2.0"                         
#>  [21] "ImageMagick/6.9.11-22"                  
#>  [22] "ImageMagick/7.0.9-5"                    
#>  [23] "JAGS/4.0.0"                             
#>  [24] "JAGS/4.3.0"                             
#>  [25] "MaxQuant/1.6.17.0"                      
#>  [26] "MaxQuant/2.0.2.0"                       
#>  [27] "ModelAngelo/0.0.1"                      
#>  [28] "MultiQC/1.10.1"                         
#>  [29] "MultiQC/gmon.out"                       
#>  [30] "NAMD-mpich/2.12"                        
#>  [31] "NAMD-mpich/2.14"                        
#>  [32] "NAMD-multicore-gpu/2.14"                
#>  [33] "NAMD-multicore/2.12"                    
#>  [34] "NAMD-multicore/2.14"                    
#>  [35] "NAMD-openmpi/2.12"                      
#>  [36] "NAMD/2.12"                              
#>  [37] "NAMD/2.13nb"                            
#>  [38] "R/3.3.3"                                
#>  [39] "R/3.4.4"                                
#>  [40] "R/3.5.2"                                
#>  [41] "R/3.5.3"                                
#>  [42] "R/3.6.0"                                
#>  [43] "R/3.6.0m"                               
#>  [44] "R/3.6.1"                                
#>  [45] "R/3.6.1m"                               
#>  [46] "R/3.6.2"                                
#>  [47] "R/3.6.3"                                
#>  [48] "R/4.0.0"                                
#>  [49] "R/4.0.1"                                
#>  [50] "R/4.0.2"                                
#>  [51] "R/4.0.3"                                
#>  [52] "R/4.0.4"                                
#>  [53] "R/4.0.5"                                
#>  [54] "R/4.1.0"                                
#>  [55] "R/4.1.1"                                
#>  [56] "R/4.1.2"                                
#>  [57] "R/4.1.3"                                
#>  [58] "R/4.2.0"                                
#>  [59] "R/4.2.1"                                
#>  [60] "R/MKL/4.2.1"                            
#>  [61] "R/openBLAS/3.6.2"                       
#>  [62] "R/openBLAS/4.2.1"                       
#>  [63] "RepeatMaker/4.1.2"                      
#>  [64] "SNPsplit/0.2.0"                         
#>  [65] "SPHIRE/1.1"                             
#>  [66] "SPHIRE/1.3"                             
#>  [67] "STAR/2.6.1c"                            
#>  [68] "STAR/2.7.0a"                            
#>  [69] "STAR/2.7.1a"                            
#>  [70] "STAR/2.7.3a"                            
#>  [71] "STAR/2.7.9a"                            
#>  [72] "TRF/4.09.1"                             
#>  [73] "Xvfb/1.19.5"                            
#>  [74] "adxv/1.9.12"                            
#>  [75] "alphafold/2.0.0"                        
#>  [76] "alphafold/2.0.1"                        
#>  [77] "alphafold/2.0.2"                        
#>  [78] "alphafold/2.1.1.0"                      
#>  [79] "alphafold/2.1.2.0"                      
#>  [80] "alphafold/2.2.0.0"                      
#>  [81] "anaconda3/2019.03"                      
#>  [82] "anaconda3/2020.07"                      
#>  [83] "anaconda3/4.0.0"                        
#>  [84] "anaconda3/4.3.1"                        
#>  [85] "anaconda3/latest"                       
#>  [86] "annovar/2015-12-14"                     
#>  [87] "apache-ant/1.9.7"                       
#>  [88] "apptainer/1.0.0"                        
#>  [89] "apptainer/1.1.0"                        
#>  [90] "aspera/3.5.4"                           
#>  [91] "aspera/3.9.1"                           
#>  [92] "aspera/3.9.6"                           
#>  [93] "autoPROC/20211020"                      
#>  [94] "autoPROC/20220608"                      
#>  [95] "autoPROC/20230217"                      
#>  [96] "autoSHARP/2.8"                          
#>  [97] "awscli/1.16py2.7"                       
#>  [98] "awscli/1.16py3.7"                       
#>  [99] "awscli/1.22.89"                         
#> [100] "awscli/2.1.25"                          
#> [101] "awscli/2.5.2"                           
#> [102] "axel/2.17.10"                           
#> [103] "bamUtil/1.0.14"                         
#> [104] "bamtools/2.4.1"                         
#> [105] "bazel/0.26.1"                           
#> [106] "bazel/1.2.1"                            
#> [107] "bcftools/1.12"                          
#> [108] "bcftools/1.13"                          
#> [109] "bcftools/1.14"                          
#> [110] "bcftools/1.15"                          
#> [111] "bcftools/1.16"                          
#> [112] "bcftools/1.17"                          
#> [113] "bcftools/1.3.1"                         
#> [114] "bcftools/1.6"                           
#> [115] "bcftools/1.7"                           
#> [116] "bcftools/1.9"                           
#> [117] "bcl-convert/3.10.5"                     
#> [118] "bcl-convert/3.9.3"                      
#> [119] "bcl2fastq/2.19.1"                       
#> [120] "bcl2fastq/2.20.0"                       
#> [121] "beast/1.8.3"                            
#> [122] "beast2/2.4.0"                           
#> [123] "bedops/2.4.26"                          
#> [124] "bedtools/2.25.0"                        
#> [125] "bedtools/2.26.0"                        
#> [126] "binutils/2.35.2-gcc-4.8.5"              
#> [127] "binutils/2.35.2-gcc-9.1.0"              
#> [128] "biobambam2/2.0.182-gcc9.1.0"            
#> [129] "bismark/0.16.1"                         
#> [130] "bismark/0.19.0"                         
#> [131] "bismark/0.19.1"                         
#> [132] "bismark/0.20.0"                         
#> [133] "boost/1.55.0"                           
#> [134] "boost/1.58.0"                           
#> [135] "boost/1.60.0"                           
#> [136] "boost/1.66.0"                           
#> [137] "bowtie/1.1.2"                           
#> [138] "bowtie2/2.2.9"                          
#> [139] "bowtie2/2.3.2"                          
#> [140] "bowtie2/2.3.4.1"                        
#> [141] "bowtie2/2.4.4"                          
#> [142] "bpipe/0.9.9.2"                          
#> [143] "breakdancer/1.1.2"                      
#> [144] "breakdancer/1.4.5"                      
#> [145] "bsoft/2.1.0"                            
#> [146] "buster/20211020"                        
#> [147] "buster/20220608"                        
#> [148] "bwa/0.7.13"                             
#> [149] "bwa/0.7.15"                             
#> [150] "bwa/0.7.17"                             
#> [151] "bwtool/1.0"                             
#> [152] "bzip2/1.0.6"                            
#> [153] "caffe2/0.8.1"                           
#> [154] "canu/1.3"                               
#> [155] "canu/1.5"                               
#> [156] "ccp4/7.1"                               
#> [157] "ccp4/8.0"                               
#> [158] "ccpem/1.5.0"                            
#> [159] "cellprofiler/4.2.1"                     
#> [160] "cellprofiler/4.2.4"                     
#> [161] "cellranger-arc/1.0.1"                   
#> [162] "cellranger-arc/2.0.2"                   
#> [163] "cellranger-atac/1.0.0"                  
#> [164] "cellranger-atac/1.2.0"                  
#> [165] "cellranger-atac/2.1.0"                  
#> [166] "cellranger/2.2.0"                       
#> [167] "cellranger/3.0.1"                       
#> [168] "cellranger/3.0.2"                       
#> [169] "cellranger/3.1.0"                       
#> [170] "cellranger/4.0.0"                       
#> [171] "cellranger/5.0.0"                       
#> [172] "cellranger/6.0.0"                       
#> [173] "cellranger/6.1.2"                       
#> [174] "cellranger/7.0.0"                       
#> [175] "cellsnp-lite/1.2.0"                     
#> [176] "chimera/1.11"                           
#> [177] "chimerax/0.8"                           
#> [178] "chimerax/0.9"                           
#> [179] "chimerax/0.93"                          
#> [180] "chimerax/1.1"                           
#> [181] "chimerax/1.3"                           
#> [182] "circos/0.69"                            
#> [183] "circos/0.69-9"                          
#> [184] "cistem/1.0.0-beta"                      
#> [185] "clara-parabricks/4.0.1-1"               
#> [186] "clover/2015-08-28"                      
#> [187] "cluster-utils/18.08.1"                  
#> [188] "cmake/3.25.1"                           
#> [189] "cnvnator/0.3.3"                         
#> [190] "computecpp/1.0.5"                       
#> [191] "coot/0.8.8"                             
#> [192] "coot/0.8.9"                             
#> [193] "coot/0.8.9.1"                           
#> [194] "crYOLO/1.7.5cpu"                        
#> [195] "crYOLO/1.7.5gpu"                        
#> [196] "crYOLO/1.8.0gpu"                        
#> [197] "crYOLO/1.9.0"                           
#> [198] "cryoEF/1.1.0"                           
#> [199] "ctffind/4.1.13"                         
#> [200] "ctffind/4.1.14"                         
#> [201] "ctffind/4.1.5"                          
#> [202] "cuda/10.0"                              
#> [203] "cuda/10.1"                              
#> [204] "cuda/10.2"                              
#> [205] "cuda/11.0"                              
#> [206] "cuda/11.1"                              
#> [207] "cuda/11.2"                              
#> [208] "cuda/11.3.1"                            
#> [209] "cuda/11.4"                              
#> [210] "cuda/11.5.1"                            
#> [211] "cuda/11.6.2"                            
#> [212] "cuda/11.7.1"                            
#> [213] "cuda/8.0"                               
#> [214] "cuda/9.2"                               
#> [215] "cudnn/10.1-7.6.5"                       
#> [216] "cudnn/11.2-8.1.1.33"                    
#> [217] "cudnn/11.3.1-8.4.0"                     
#> [218] "cudnn/11.3.1-8.6.0.162"                 
#> [219] "curl/7.65.0"                            
#> [220] "cutadapt/2.9"                           
#> [221] "cutadapt/3.4"                           
#> [222] "cutadapt/3.7"                           
#> [223] "dalilite/5.1"                           
#> [224] "darknet/61c9d02cu10.0"                  
#> [225] "deepEMhancer/0.15"                      
#> [226] "deeptools/3.5.1"                        
#> [227] "depot_tools/6c7b829"                    
#> [228] "dials/3.10.0"                           
#> [229] "dials/3.12.0"                           
#> [230] "dorado/0.1.1-gcc-9.1.0"                 
#> [231] "dotnet/2.1.809"                         
#> [232] "dotnet/3.1.412"                         
#> [233] "doublecmd/0.9.10.gtk2"                  
#> [234] "dua-cli/2.17.8"                         
#> [235] "dua-cli/2.19.0"                         
#> [236] "duck/6.9.3"                             
#> [237] "duckdb/0.6.1"                           
#> [238] "edirect/15.0"                           
#> [239] "egaclient/2.2.2"                        
#> [240] "eiger2cbf/160415"                       
#> [241] "elbencho/1.7-1cu10.1"                   
#> [242] "ensembl-vep/89.0"                       
#> [243] "ensembl-vep/89.8"                       
#> [244] "ensembl-vep/98.3"                       
#> [245] "evince/3.28.2"                          
#> [246] "fastqc/0.11.5"                          
#> [247] "fastqc/0.11.8"                          
#> [248] "fastx-toolkit/0.0.13"                   
#> [249] "feh/3.6.3"                              
#> [250] "ffmpeg/3.4"                             
#> [251] "fftw/3.3.9"                             
#> [252] "fio/3.16"                               
#> [253] "fio/3.19"                               
#> [254] "flash/1.2.11"                           
#> [255] "fluxbox/1.3.7"                          
#> [256] "freebayes/1.1.0"                        
#> [257] "freebayes/1.2.0"                        
#> [258] "gatk/4.2.4.1"                           
#> [259] "gatk/4.2.5.0"                           
#> [260] "gautomatch/0.56cu10.1"                  
#> [261] "gautomatch/0.56cu8.0"                   
#> [262] "gcc/10.3.0"                             
#> [263] "gcc/11.1.0"                             
#> [264] "gcc/11.2.0"                             
#> [265] "gcc/6.3.0"                              
#> [266] "gcc/7.1.0"                              
#> [267] "gcc/7.3.0"                              
#> [268] "gcc/8.1.0"                              
#> [269] "gcc/8.2.0"                              
#> [270] "gcc/8.3.0"                              
#> [271] "gcc/9.1.0"                              
#> [272] "gdal/2.4.4"                             
#> [273] "gdal/3.6.2"                             
#> [274] "gem/3.0"                                
#> [275] "gemini/0.20.1"                          
#> [276] "gencore/0.17.2"                         
#> [277] "genetorrent/3.8.7"                      
#> [278] "genozip/12.0.11"                        
#> [279] "geos/3.7.0"                             
#> [280] "germline/1.5.1"                         
#> [281] "gh-cli/2.0"                             
#> [282] "git/2.37.1"                             
#> [283] "glew/1.9.0"                             
#> [284] "glib/1.2.10"                            
#> [285] "gmap-gsnap/2017-05-08"                  
#> [286] "gmap-gsnap/2017-06-16"                  
#> [287] "gnuplot/5.2.6"                          
#> [288] "go/1.19.4"                              
#> [289] "google-cloud-sdk/194.0.0"               
#> [290] "google-cloud-sdk/348.0.0"               
#> [291] "gradle/4.3"                             
#> [292] "graphmap/0.3.1"                         
#> [293] "graphmap/0.5.1"                         
#> [294] "graphmap/0.5.2"                         
#> [295] "gridss/2.11.1"                          
#> [296] "gromacs/2021.2"                         
#> [297] "gromacs/4.5.5_cryo_fit"                 
#> [298] "groovy/4.0.0"                           
#> [299] "gsl/2.4"                                
#> [300] "gsl/2.6"                                
#> [301] "guppy-cpu/2.3.7"                        
#> [302] "guppy-cpu/3.1.5"                        
#> [303] "guppy-cpu/3.2.1"                        
#> [304] "guppy-cpu/3.2.2"                        
#> [305] "guppy-cpu/3.2.4"                        
#> [306] "guppy-cpu/3.3.0"                        
#> [307] "guppy-cpu/3.6.1"                        
#> [308] "guppy-cpu/4.0.15"                       
#> [309] "guppy-gpu/2.3.1"                        
#> [310] "guppy-gpu/2.3.7"                        
#> [311] "guppy-gpu/3.1.5"                        
#> [312] "guppy-gpu/3.2.1"                        
#> [313] "guppy-gpu/3.2.2"                        
#> [314] "guppy-gpu/3.2.4"                        
#> [315] "guppy-gpu/3.3.0"                        
#> [316] "guppy-gpu/3.6.1"                        
#> [317] "guppy-gpu/4.0.15"                       
#> [318] "guppy-gpu/6.2.1"                        
#> [319] "gzip/1.10"                              
#> [320] "haplopainter/1.043"                     
#> [321] "hdf4/4.2.11"                            
#> [322] "hdf5-mpich/1.10.5_3.3"                  
#> [323] "hdf5/1.10.5"                            
#> [324] "hdf5/1.12.1"                            
#> [325] "hdf5/1.12.2"                            
#> [326] "hdf5/1.8.16"                            
#> [327] "hdf5/1.8.20"                            
#> [328] "hdf5/1.8.21"                            
#> [329] "hdfview/2.11.0"                         
#> [330] "hisat2/2.0.3"                           
#> [331] "hisat2/2.0.5"                           
#> [332] "hmmer/2.3.2"                            
#> [333] "hmmer/3.1b2"                            
#> [334] "hmmer/3.1b2-mpi"                        
#> [335] "hmmer/3.2.1"                            
#> [336] "hmmer/3.3.2"                            
#> [337] "homer/4.11"                             
#> [338] "hpl/2.3"                                
#> [339] "htop/2.0.2"                             
#> [340] "htslib/1.10.2"                          
#> [341] "htslib/1.12"                            
#> [342] "htslib/1.13"                            
#> [343] "htslib/1.14"                            
#> [344] "htslib/1.15"                            
#> [345] "htslib/1.17"                            
#> [346] "htslib/1.3.1"                           
#> [347] "htslib/1.3.2"                           
#> [348] "htslib/1.4.1"                           
#> [349] "htslib/1.5"                             
#> [350] "htslib/1.6"                             
#> [351] "htslib/1.7"                             
#> [352] "htslib/1.9"                             
#> [353] "icewm/2.8.0"                            
#> [354] "iftop/1.0"                              
#> [355] "imp/2.5.0"                              
#> [356] "intel-ipp/2019.5.281"                   
#> [357] "intel-mkl/2019.3.199"                   
#> [358] "intel-mpi/2019.3.199"                   
#> [359] "intel-tbb/2019.3.199"                   
#> [360] "intel_mkl_2019/2019.5.075"              
#> [361] "intel_mpi_2019/2019.5.075"              
#> [362] "intel_mpi_2019/default_pipeline.star"   
#> [363] "ior-slurm/3.2.1mpich3.3"                
#> [364] "ior-slurm/3.2.1openMPI4.0.2"            
#> [365] "ior/3.2.1mpio4.0.1"                     
#> [366] "iozone/3.491"                           
#> [367] "java/1.7.0_79"                          
#> [368] "java/1.8.0_131"                         
#> [369] "java/1.8.0_151"                         
#> [370] "java/1.8.0_211"                         
#> [371] "java/1.8.0_92"                          
#> [372] "jbrowse-desktop/2.3.2"                  
#> [373] "jellyfish/1.1.11"                       
#> [374] "jellyfish/2.2.10"                       
#> [375] "joe/4.6"                                
#> [376] "joe/n"                                  
#> [377] "julia/0.6.4"                            
#> [378] "julia/1.0.1"                            
#> [379] "julia/1.5.3"                            
#> [380] "julia/1.8.5"                            
#> [381] "juncBASE/1.2-beta"                      
#> [382] "jupyter-config/1.0"                     
#> [383] "kallisto/0.44.0"                        
#> [384] "kallisto/0.46.0"                        
#> [385] "kraken/2.1.1"                           
#> [386] "lapack/3.8.0"                           
#> [387] "last/864"                               
#> [388] "libaio/0.3.111"                         
#> [389] "libglvnd/1.1.1"                         
#> [390] "libiconv/1.16"                          
#> [391] "libmaus2/2.0.798-gcc9.1.0"              
#> [392] "libtorch-gpu/1.12.1"                    
#> [393] "lighttpd/1.4.45"                        
#> [394] "lxrandr/0.3.1"                          
#> [395] "lz4/1.9.3"                              
#> [396] "macs2/2.2.7.1"                          
#> [397] "mageck/0.5.9.4"                         
#> [398] "magma-gpu/2.7.0-11.3.1"                 
#> [399] "mariadb-client/10.11.2"                 
#> [400] "mariadb-connector-c/3.1.11"             
#> [401] "mariadb-connector-c/3.3.4"              
#> [402] "maven/3.3.9"                            
#> [403] "maven/3.5.0"                            
#> [404] "mbuffer/20171011"                       
#> [405] "mcl/14-137"                             
#> [406] "mediaflux-data-mover/1.0.11"            
#> [407] "mediaflux-data-mover/1.0.12"            
#> [408] "mediaflux-data-mover/1.0.8"             
#> [409] "mediaflux-data-mover/1.1.15"            
#> [410] "meme/4.12.0"                            
#> [411] "meme/5.0.5"                             
#> [412] "meme/openmpi-4.12.0"                    
#> [413] "meme/openmpi-5.0.5"                     
#> [414] "mercurial/4.8"                          
#> [415] "miniconda2/4.5.12"                      
#> [416] "miniconda3-alt/4.10.3"                  
#> [417] "miniconda3/4.10.3"                      
#> [418] "minimap2/2.17"                          
#> [419] "minimap2/2.4"                           
#> [420] "molgenis-compute/17.08.1"               
#> [421] "mono/6.10.0.104"                        
#> [422] "mono/6.12.0.90"                         
#> [423] "motioncor2/1.2.1cu8.0"                  
#> [424] "motioncor2/1.2.1cu9.2"                  
#> [425] "motioncor2/1.2.2cu10.0"                 
#> [426] "motioncor2/1.2.2cu8.0"                  
#> [427] "motioncor2/1.2.2cu9.2"                  
#> [428] "motioncor2/1.2.6cu10.1"                 
#> [429] "motioncor2/1.2.6cu8.0"                  
#> [430] "motioncor2/1.2.6cu9.2"                  
#> [431] "motioncor2/1.3.1cu10.1"                 
#> [432] "motioncor2/1.3.1cu9.2"                  
#> [433] "motioncor2/1.4.0cu10.0"                 
#> [434] "motioncor2/1.4.0cu10.1"                 
#> [435] "motioncor2/1.4.0cu10.2"                 
#> [436] "motioncor2/1.4.0cu11.0"                 
#> [437] "motioncor2/1.4.0cu9.2"                  
#> [438] "motioncor2/1.4.4cu10.0"                 
#> [439] "motioncor2/1.4.4cu10.1"                 
#> [440] "motioncor2/1.4.4cu10.2"                 
#> [441] "motioncor2/1.4.4cu11.1"                 
#> [442] "motioncor2/1.4.4cu11.2"                 
#> [443] "motioncor2/1.4.4cu11.3.1"               
#> [444] "motioncor2/1.5.0cu10.1"                 
#> [445] "motioncor2/1.5.0cu10.2"                 
#> [446] "motioncor2/1.5.0cu11.1"                 
#> [447] "motioncor2/1.5.0cu11.2"                 
#> [448] "motioncor2/1.5.0cu11.4"                 
#> [449] "motioncor2/11.5.0cu11.3.1"              
#> [450] "motioncorr/2.1"                         
#> [451] "mpfr/4.1.0"                             
#> [452] "mpich-slurm/3.3.2"                      
#> [453] "mpich-slurm/3.4.1"                      
#> [454] "mpich-slurm/3.4.2"                      
#> [455] "mpich/3.3"                              
#> [456] "mpich/3.3.2"                            
#> [457] "mrcz/0.1.3"                             
#> [458] "multiSNV/2.3-10-g06102a5"               
#> [459] "muscle/3.8.31"                          
#> [460] "mutect/1.1.7"                           
#> [461] "nanonet-openBLAS/2.0.0"                 
#> [462] "nanonet/2.0.0"                          
#> [463] "nanopolish/0.5.0"                       
#> [464] "nanopolish/0.6.2"                       
#> [465] "nanopolish/0.7.1"                       
#> [466] "nanopolish/0.8.1"                       
#> [467] "nanopolish/0.8.5"                       
#> [468] "nanopore/1.0"                           
#> [469] "ncbi-blast/2.10.1+"                     
#> [470] "ncbi-blast/2.11.0+"                     
#> [471] "ncbi-blast/2.6.0+"                      
#> [472] "ncbi-blast/2.7.1+"                      
#> [473] "ncbi-blast/2.9.0"                       
#> [474] "ncbi-blast/2.9.0a"                      
#> [475] "ncftp/3.2.6"                            
#> [476] "nedit/5.5"                              
#> [477] "netcdf/4.4.1.1"                         
#> [478] "nextflow/22.04.5"                       
#> [479] "nextflow/22.10.4"                       
#> [480] "nf-core/2.7.2"                          
#> [481] "ninja/1.10.0"                           
#> [482] "nix/fix"                                
#> [483] "nmap-ncat/7.91"                         
#> [484] "node/6.11.0"                            
#> [485] "node/8.11.2"                            
#> [486] "node/9.9.0"                             
#> [487] "nodejs/10.24.1"                         
#> [488] "nodejs/16.19.0"                         
#> [489] "nodejs/17.9.1"                          
#> [490] "novoalign/3.04.04"                      
#> [491] "novocraft/3.08.02"                      
#> [492] "oases/0.2.09"                           
#> [493] "ocl-icd/2.3.1"                          
#> [494] "octave/6.4.0-gcc11.1.0"                 
#> [495] "oneMKL/2022.1.0.223"                    
#> [496] "openBLAS/0.3.21-gcc-11.1.0"             
#> [497] "openBLAS/0.3.21-gcc-11.1.0-skylake"     
#> [498] "openBLAS/0.3.6-gcc-9.1.0"               
#> [499] "openCV/2.4.13.6"                        
#> [500] "openCV/4.2.0"                           
#> [501] "openMPI-slurm/4.1.0"                    
#> [502] "openMPI-slurm/4.1.2"                    
#> [503] "openMPI/4.1.4"                          
#> [504] "openSSL/1.0.2r"                         
#> [505] "openSSL/1.1.1b"                         
#> [506] "openSSL/1.1.1g"                         
#> [507] "openSSL/1.1.1k"                         
#> [508] "openSSL/1.1.1n"                         
#> [509] "openjdk/1.8.0"                          
#> [510] "openjdk/13.0.2"                         
#> [511] "openjdk/14.0.2"                         
#> [512] "openjdk/15.0.2"                         
#> [513] "openjdk/16.0.1"                         
#> [514] "openjdk/17.0.2"                         
#> [515] "openjdk/18.0.2"                         
#> [516] "openmpi-CUDA8/3.0.0"                    
#> [517] "openmpi/2.1.0"                          
#> [518] "openmpi/2.1.1"                          
#> [519] "openmpi/3.0.0"                          
#> [520] "openmpi/4.1.1"                          
#> [521] "openmpi/4.1.1-slurm"                    
#> [522] "owncloud-client/2.3.3"                  
#> [523] "pandoc/1.17.0.3"                        
#> [524] "pandoc/2.0.0.1"                         
#> [525] "pandoc/2.14.2"                          
#> [526] "pandoc/2.19.2"                          
#> [527] "pandoc/2.3.1"                           
#> [528] "parallel/20190722"                      
#> [529] "pbs-drmaa/1.0.19"                       
#> [530] "pbstop/1.0"                             
#> [531] "pcre2/10.34-gcc-9.1.0"                  
#> [532] "pcre2/10.36-gcc-11.1.0"                 
#> [533] "peakseq/1.31"                           
#> [534] "pear/0.9.10"                            
#> [535] "pear/0.9.11"                            
#> [536] "perl/5.22.2"                            
#> [537] "perl/5.26.3"                            
#> [538] "perl/5.30.1"                            
#> [539] "pgsql/15.1"                             
#> [540] "phenix-nightly/1.15-3459"               
#> [541] "phenix-rosetta/1.11.1-2575"             
#> [542] "phenix/1.15-3459"                       
#> [543] "phenix/1.17-3644"                       
#> [544] "phenix/1.18.2-3874"                     
#> [545] "phenix/1.20.1-4487"                     
#> [546] "php/5.6.21"                             
#> [547] "picard-tools/2.26.11"                   
#> [548] "pigz/2.6"                               
#> [549] "pindel/0.2.5b9"                         
#> [550] "pixz/1.0.6"                             
#> [551] "platypus/0.8.1"                         
#> [552] "plink/1.9"                              
#> [553] "pmix/2.2.5"                             
#> [554] "pmix/3.2.3"                             
#> [555] "poretools/0.6.0"                        
#> [556] "proj/4.9.3"                             
#> [557] "proj/6.3.2"                             
#> [558] "proj/9.1.0"                             
#> [559] "pyega3/pyega3-3.4.0/bin/pyega3"         
#> [560] "pyega3/pyega3-3.4.0/pyega3-3.4.0.sif"   
#> [561] "pyem/0.5"                               
#> [562] "pymol/1.7.4.0"                          
#> [563] "pymol/2.5.0"                            
#> [564] "python/2.7.18"                          
#> [565] "python/3.10.4"                          
#> [566] "python/3.5.1"                           
#> [567] "python/3.5.3"                           
#> [568] "python/3.6.5-intel"                     
#> [569] "python/3.7.0"                           
#> [570] "python/3.7.13"                          
#> [571] "python/3.8.3"                           
#> [572] "python/3.8.8"                           
#> [573] "python/3.9.5"                           
#> [574] "qiime2/2019.7"                          
#> [575] "qiime2/2020.8"                          
#> [576] "qpdf/10.0.1"                            
#> [577] "quarto/1.1.189"                         
#> [578] "rc-tools/1.0"                           
#> [579] "rclone/1.55.0"                          
#> [580] "relion-cpu/3.1b-icc"                    
#> [581] "relion-cpu/3.1beta-icc"                 
#> [582] "relion/3.1.3-cu11.2"                    
#> [583] "relion/3.1.3-gcc11.1.0-cu11.4"          
#> [584] "relion/4.0-beta-1-cu10.1"               
#> [585] "relion/4.0-beta-1-cu11.2"               
#> [586] "relion/4.0-cu11.3.1"                    
#> [587] "rmblast/2.11.0"                         
#> [588] "rnammer/1.2"                            
#> [589] "root/6.10.00"                           
#> [590] "rsem/1.3.0"                             
#> [591] "rsem/1.3.1"                             
#> [592] "rstudio_singularity/1.0.0"              
#> [593] "ruby/2.3.1"                             
#> [594] "rust/1.63.0"                            
#> [595] "salmon/0.11.2"                          
#> [596] "salmon/0.14.1"                          
#> [597] "salmon/0.9.1"                           
#> [598] "salmon/1.1.0"                           
#> [599] "salmon/1.5.2"                           
#> [600] "salmon/1.9.0"                           
#> [601] "salmonTE/0.3"                           
#> [602] "sambamba/0.6.6"                         
#> [603] "samblaster/0.1.23"                      
#> [604] "samtools/0.1.18"                        
#> [605] "samtools/0.1.19"                        
#> [606] "samtools/1.10"                          
#> [607] "samtools/1.12"                          
#> [608] "samtools/1.13"                          
#> [609] "samtools/1.14"                          
#> [610] "samtools/1.15"                          
#> [611] "samtools/1.16.1"                        
#> [612] "samtools/1.17"                          
#> [613] "samtools/1.3.1"                         
#> [614] "samtools/1.4.1"                         
#> [615] "samtools/1.5"                           
#> [616] "samtools/1.6"                           
#> [617] "samtools/1.7"                           
#> [618] "samtools/1.9"                           
#> [619] "scala/2.12.2"                           
#> [620] "schrodinger/2022-1"                     
#> [621] "schrodinger/2022-2"                     
#> [622] "schrodinger/2022-3"                     
#> [623] "scipion/2.0"                            
#> [624] "scipion/3.0"                            
#> [625] "seqmonk/1.45.1"                         
#> [626] "seqmonk/1.46"                           
#> [627] "seqmonk/1.47.0"                         
#> [628] "seqmonk/1.47.1"                         
#> [629] "seqmonk/1.47.2"                         
#> [630] "seqmonk/1.48.0"                         
#> [631] "seqmonk/1.48.1"                         
#> [632] "sidesplitter/1.2"                       
#> [633] "signalp/4.1"                            
#> [634] "singularity/2.5.1"                      
#> [635] "singularity/3.0.3"                      
#> [636] "singularity/3.1.1"                      
#> [637] "singularity/3.3.0"                      
#> [638] "singularity/3.5.0"                      
#> [639] "singularity/3.5.2"                      
#> [640] "singularity/3.6.2"                      
#> [641] "singularity/3.7.4"                      
#> [642] "singularity/3.8.2"                      
#> [643] "singularity/3.8.4"                      
#> [644] "slurm-contribs/20.11.5"                 
#> [645] "snakemake/7.12.0"                       
#> [646] "snpEff/4.3i"                            
#> [647] "snpEff/4.3t"                            
#> [648] "soapdenovo-trans/1.03"                  
#> [649] "soapdenovo2/2.04"                       
#> [650] "spaceranger/1.2.2"                      
#> [651] "spaceranger/1.3.0"                      
#> [652] "spaceranger/2.0.0"                      
#> [653] "spades/3.11.0"                          
#> [654] "spawn-fcgi/1.6.5"                       
#> [655] "spp/1.14"                               
#> [656] "sqlite/3.38.5"                          
#> [657] "sqlite/3.40.0"                          
#> [658] "sqlite/3.40.1"                          
#> [659] "sra-toolkit/2.10.9"                     
#> [660] "sra-toolkit/2.8.2-1"                    
#> [661] "sra-toolkit/2.9.6-1"                    
#> [662] "sra-toolkit/3.0.0"                      
#> [663] "sshpass/1.05"                           
#> [664] "stack/1.5.1"                            
#> [665] "stack/1.7.1"                            
#> [666] "stornext/1.0"                           
#> [667] "stubl/0.0.10"                           
#> [668] "subread/1.5.0-p2"                       
#> [669] "subread/1.5.1"                          
#> [670] "subread/1.5.2"                          
#> [671] "subread/1.6.3"                          
#> [672] "summovie/1.0.2"                         
#> [673] "sxgui_settings/gui_settings_project.txt"
#> [674] "tadbit/0.2"                             
#> [675] "tadbit/gmon.out"                        
#> [676] "tar/1.34"                               
#> [677] "tcpdump/4.9.2"                          
#> [678] "tensorRT/10.0-6.0.1.5"                  
#> [679] "tensorRT/10.1-6.0.1.5"                  
#> [680] "tensorRT/11.2-7.2.3.4"                  
#> [681] "tensorflow-gpu/10.0-1.15.2"             
#> [682] "tensorflow-gpu/10.1-2.1.0"              
#> [683] "tensorflow-gpu/11.2-2.11.0"             
#> [684] "texlive/20171130"                       
#> [685] "texlive/3.14159265"                     
#> [686] "tmhmm/2.0c"                             
#> [687] "topaz/0.2.3"                            
#> [688] "topaz/0.2.4"                            
#> [689] "tophat/2.1.1"                           
#> [690] "transdecoder/4.0.0"                     
#> [691] "trickle/1.07"                           
#> [692] "trimgalore/0.4.2"                       
#> [693] "trimgalore/0.4.4"                       
#> [694] "trimgalore/0.6.7"                       
#> [695] "trimmomatic/0.36"                       
#> [696] "trinityrnaseq/2.2.0"                    
#> [697] "trinityrnaseq/2.4.0"                    
#> [698] "trinityrnaseq/2.8.4"                    
#> [699] "trinotate/3.0.1"                        
#> [700] "trinotate/3.0.2"                        
#> [701] "turbovnc/3.0.1"                         
#> [702] "tvips/0.0.3"                            
#> [703] "ucsc-tools/331"                         
#> [704] "ultraplex/1.2.5"                        
#> [705] "umi_tools/1.0.1"                        
#> [706] "unblur/1.0.2"                           
#> [707] "usearch/7.0.1090"                       
#> [708] "usearch/8.1.1861"                       
#> [709] "usearch/9.2.64"                         
#> [710] "v8/3.14.5.10"                           
#> [711] "vardict/1.5.1"                          
#> [712] "varscan/2.3.9"                          
#> [713] "vcflib/1.0.0-rc1"                       
#> [714] "vcftools/0.1.13"                        
#> [715] "velvet/1.2.10"                          
#> [716] "vireoSNP/0.5.6"                         
#> [717] "vsearch/2.4.3"                          
#> [718] "websockify/0.9.0"                       
#> [719] "wehi-stornext/1.0"                      
#> [720] "wine/6.9"                               
#> [721] "wine/7.17"                              
#> [722] "xds/20220110"                           
#> [723] "xdsgui/2017-01-25"                      
#> [724] "xdsgui/2020-12-02"                      
#> [725] "xenomapper/1.0.1"                       
#> [726] "xz/5.2.2"                               
#> [727] "zeromq/4.3.4"                           
#> [728] "zlib/1.2.11"                            
#> [729] "zlib/1.2.8"                             
#> [730] "zstd/1.5.0"