Lists all modules available to be loaded

Lists all modules available to be loaded

Usage

module_avail(starts_with = NULL, contains = NULL)

Arguments

starts_with

An optional character scalar. If provided, only modules whose name starts with this character string will be returned.

contains

An optional character scalar. This parameter is only supported in Environment Modules 4.3+. If provided, it will filter modules to only those containing this substring.

Value

A character vector with class "cli_ansi_string".

Examples

module_avail()
#> <cli_ansi_string>
#> [1] ------------------- /stornext/System/data/modulefiles/tools --------------------                          
#> [2] apptainer/1.0.0              mpich-slurm/3.3.2                   wine/6.9        
#> [3] apptainer/1.1.0              mpich-slurm/3.4.1                   wine/7.17                                                 
#> [4] aspera/3.5.4                 mpich-slurm/3.4.2                   zeromq/4.3.4                                
#> [5] aspera/3.9.1                 mpich/3.3                           zstd/1.5.0                    
#> [6] aspera/3.9.6                 mpich/3.3.2                                                                     
#> [7] awscli/1.16py2.7             ncftp/3.2.6                                                                     
#> [8] awscli/1.16py3.7             nextflow/22.04.5                                                                              
#> [9] awscli/1.22.89               nextflow/22.10.4                                                                
#> [10] awscli/2.1.25                ninja/1.10.0                                                                   
#> [11] awscli/2.5.2                 nmap-ncat/7.91                                                   
#> [12] axel/2.17.10                 nodejs/10.24.1                                                                 
#> [13] bazel/0.26.1                 nodejs/16.19.0                                                                               
#> [14] bazel/1.2.1                  nodejs/17.9.1                                                    
#> [15] binutils/2.35.2-gcc-4.8.5    ocl-icd/2.3.1                                                                  
#> [16] binutils/2.35.2-gcc-9.1.0    octave/6.4.0-gcc11.1.0                                                         
#> [17] cluster-utils/18.08.1        oneMKL/2022.1.0.223                                              
#> [18] cmake/3.25.1                 openBLAS/0.3.6-gcc-9.1.0                                         
#> [19] CUnit/2.1-3                  openBLAS/0.3.21-gcc-11.1.0                                                     
#> [20] curl/7.65.0                  openBLAS/0.3.21-gcc-11.1.0-skylake                                             
#> [21] depot_tools/6c7b829          openCV/2.4.13.6                                                                
#> [22] doublecmd/0.9.10.gtk2        openCV/4.2.0                                                     
#> [23] dua-cli/2.17.8               openjdk/1.8.0                                                                  
#> [24] dua-cli/2.19.0               openjdk/13.0.2                                                                               
#> [25] elbencho/1.7-1cu10.1         openjdk/14.0.2                                                                               
#> [26] evince/3.28.2                openjdk/15.0.2                                                                 
#> [27] feh/3.6.3                    openjdk/16.0.1                                                                 
#> [28] fftw/3.3.9                   openjdk/17.0.2                                                                 
#> [29] fio/3.16                     openjdk/18.0.2                                                                 
#> [30] fio/3.19                     openMPI-slurm/4.1.0                                                            
#> [31] fluxbox/1.3.7                openMPI-slurm/4.1.2                                              
#> [32] gdal/2.4.4                   openmpi/2.1.0                                                                  
#> [33] gdal/3.6.2                   openmpi/2.1.1                                                                                
#> [34] git/2.37.1                   openmpi/3.0.0                                                                  
#> [35] gnuplot/5.2.6                openmpi/4.1.1                                                    
#> [36] go/1.19.4                    openmpi/4.1.1-slurm                                                            
#> [37] groovy/4.0.0                 openMPI/4.1.4                                                    
#> [38] gzip/1.10                    openSSL/1.0.2r                                                                 
#> [39] hdf5-mpich/1.10.5_3.3        openSSL/1.1.1b                                                                 
#> [40] hpl/2.3                      openSSL/1.1.1g                                                                 
#> [41] icewm/2.8.0                  openSSL/1.1.1k                                                                 
#> [42] iftop/1.0                    openSSL/1.1.1n                                                   
#> [43] ImageMagick/6.9.11-22        owncloud-client/2.3.3                                                          
#> [44] ImageMagick/7.0.9-5          pandoc/2.14.2                                                                  
#> [45] intel-ipp/2019.5.281         pandoc/2.19.2                                                    
#> [46] intel-mkl/2019.3.199         pgsql/15.1                                                       
#> [47] intel-mpi/2019.3.199         pigz/2.6                                                         
#> [48] intel-tbb/2019.3.199         pmix/2.2.5                                                                     
#> [49] intel_mkl_2019/2019.5.075    pmix/3.2.3                                                                     
#> [50] intel_mpi_2019/2019.5.075    proj/4.9.3                                                                     
#> [51] ior-slurm/3.2.1mpich3.3      proj/6.3.2                                                                     
#> [52] ior-slurm/3.2.1openMPI4.0.2  proj/9.1.0                                                                                   
#> [53] ior/3.2.1mpio4.0.1           qpdf/10.0.1                                                      
#> [54] iozone/3.491                 quarto/1.1.189                                                                 
#> [55] julia/0.6.4                  rclone/1.55.0                                                                  
#> [56] julia/1.0.1                  rstudio_singularity/1.0.0                                                      
#> [57] julia/1.5.3                  rust/1.63.0                                                                    
#> [58] julia/1.8.5                  slurm-contribs/20.11.5                                           
#> [59] libaio/0.3.111               snakemake/7.12.0                                                 
#> [60] libiconv/1.16                sqlite/3.38.5                                                                  
#> [61] lz4/1.9.3                    sqlite/3.40.0                                                                  
#> [62] mariadb-connector-c/3.1.11   sqlite/3.40.1                                                    
#> [63] maven/3.3.9                  stubl/0.0.10                                                     
#> [64] mediaflux-data-mover/1.0.8   tar/1.34                                                                       
#> [65] mediaflux-data-mover/1.0.11  tcpdump/4.9.2                                                                  
#> [66] mediaflux-data-mover/1.0.12  trickle/1.07                                                     
#> [67] mediaflux-data-mover/1.1.15  v8/3.14.5.10                                                                   
#> [68] mpfr/4.1.0                   websockify/0.9.0                                                 
#> [69]                                                                                                                           
#> [70] ----------------- /stornext/System/data/modulefiles/bioinf/its -----------------                         
#> [71] 7zip/15.14.1                 hdf5/1.8.20                R/3.6.0                                             
#> [72] anaconda3/4.0.0              hdf5/1.8.21                R/3.6.0m                                                          
#> [73] anaconda3/4.3.1              hdf5/1.10.5                R/3.6.1                                                           
#> [74] anaconda3/2019.03            hdf5/1.12.1                R/3.6.1m                              
#> [75] anaconda3/2020.07            hdf5/1.12.2                R/3.6.2                                                           
#> [76] anaconda3/latest             hdfview/2.11.0             R/3.6.3                                             
#> [77] annovar/2015-12-14           HiC-Pro/2.11.4             R/4.0.0                               
#> [78] apache-ant/1.9.7             hisat2/2.0.3               R/4.0.1                                             
#> [79] bamtools/2.4.1               hisat2/2.0.5               R/4.0.2                               
#> [80] bamUtil/1.0.14               hmmer/2.3.2                R/4.0.3                                             
#> [81] bcftools/1.3.1               hmmer/3.1b2                R/4.0.4                                                           
#> [82] bcftools/1.6                 hmmer/3.1b2-mpi            R/4.0.5                                                           
#> [83] bcftools/1.7                 hmmer/3.2.1                R/4.1.0                                                           
#> [84] bcftools/1.9                 hmmer/3.3.2                R/4.1.1                                             
#> [85] bcftools/1.12                homer/4.11                 R/4.1.2                                             
#> [86] bcftools/1.13                htop/2.0.2                 R/4.1.3                                             
#> [87] bcftools/1.14                htslib/1.3.1               R/4.2.0                                                           
#> [88] bcftools/1.15                htslib/1.3.2               R/4.2.1                                                           
#> [89] bcftools/1.16                htslib/1.4.1               R/MKL/4.2.1                           
#> [90] bcl-convert/3.9.3            htslib/1.5                 R/openBLAS/3.6.2                                                  
#> [91] bcl-convert/3.10.5           htslib/1.6                 R/openBLAS/4.2.1                      
#> [92] bcl2fastq/2.19.1             htslib/1.7                 rc-tools/1.0                          
#> [93] bcl2fastq/2.20.0             htslib/1.9                 RepeatMaker/4.1.2                                   
#> [94] beast/1.8.3                  htslib/1.10.2              rmblast/2.11.0                        
#> [95] beast2/2.4.0                 htslib/1.12                rnammer/1.2                           
#> [96] bedops/2.4.26                htslib/1.13                root/6.10.00                          
#> [97] bedtools/2.25.0              htslib/1.14                rsem/1.3.0                                                        
#> [98] bedtools/2.26.0              htslib/1.15                rsem/1.3.1              
#> [99] biobambam2/2.0.182-gcc9.1.0  IGV/2.5.2                  ruby/2.3.1                            
#> [100] bismark/0.16.1               IGV/2.10.3                 salmon/0.9.1                                       
#> [101] bismark/0.19.0               imp/2.5.0                  salmon/0.11.2                                      
#> [102] bismark/0.19.1               IRFinder/1.2.0             salmon/0.14.1                                      
#> [103] bismark/0.20.0               JAGS/4.0.0                 salmon/1.1.0                                       
#> [104] boost/1.55.0                 JAGS/4.3.0                 salmon/1.5.2                                       
#> [105] boost/1.58.0                 java/1.7.0_79              salmon/1.9.0                                       
#> [106] boost/1.60.0                 java/1.8.0_92              salmonTE/0.3                         
#> [107] boost/1.66.0                 java/1.8.0_131             sambamba/0.6.6                                     
#> [108] bowtie/1.1.2                 java/1.8.0_151             samblaster/0.1.23                    
#> [109] bowtie2/2.2.9                java/1.8.0_211             samtools/0.1.18                                    
#> [110] bowtie2/2.3.2                jbrowse-desktop/2.3.2      samtools/0.1.19                                    
#> [111] bowtie2/2.3.4.1              jellyfish/1.1.11           samtools/1.3.1                                                   
#> [112] bowtie2/2.4.4                jellyfish/2.2.10           samtools/1.4.1                       
#> [113] bpipe/0.9.9.2                joe/4.6                    samtools/1.5                         
#> [114] breakdancer/1.1.2            juncBASE/1.2-beta          samtools/1.6                                       
#> [115] breakdancer/1.4.5            jupyter-config/1.0         samtools/1.7                         
#> [116] bwa/0.7.13                   kallisto/0.44.0            samtools/1.9                                                     
#> [117] bwa/0.7.15                   kallisto/0.46.0            samtools/1.10                                      
#> [118] bwa/0.7.17                   kraken/2.1.1               samtools/1.12                        
#> [119] bwtool/1.0                   lapack/3.8.0               samtools/1.13                        
#> [120] bzip2/1.0.6                  last/864                   samtools/1.14                        
#> [121] caffe2/0.8.1                 libmaus2/2.0.798-gcc9.1.0  samtools/1.15                        
#> [122] canu/1.3                     lighttpd/1.4.45            samtools/1.16.1        
#> [123] canu/1.5                     lxrandr/0.3.1              scala/2.12.2                         
#> [124] cellprofiler/4.2.1           macs2/2.2.7.1              seqmonk/1.45.1                                     
#> [125] cellprofiler/4.2.4           mageck/0.5.9.4             seqmonk/1.46                         
#> [126] cellranger-arc/1.0.1         maven/3.5.0                seqmonk/1.47.0                       
#> [127] cellranger-arc/2.0.2         MaxQuant/1.6.17.0          seqmonk/1.47.1                                                   
#> [128] cellranger-atac/1.0.0        MaxQuant/2.0.2.0           seqmonk/1.47.2                       
#> [129] cellranger-atac/1.2.0        mbuffer/20171011           seqmonk/1.48.0                                     
#> [130] cellranger-atac/2.1.0        mcl/14-137                 seqmonk/1.48.1                       
#> [131] cellranger/2.2.0             meme/4.12.0                signalp/4.1                                        
#> [132] cellranger/3.0.1             meme/5.0.5                 singularity/2.5.1                                  
#> [133] cellranger/3.0.2             meme/openmpi-4.12.0        singularity/3.0.3                                                
#> [134] cellranger/3.1.0             meme/openmpi-5.0.5         singularity/3.1.1                                                
#> [135] cellranger/4.0.0             mercurial/4.8              singularity/3.3.0                                  
#> [136] cellranger/5.0.0             miniconda2/4.5.12          singularity/3.5.0                    
#> [137] cellranger/6.0.0             miniconda3-alt/4.10.3      singularity/3.5.2                                  
#> [138] cellranger/6.1.2             miniconda3/4.10.3          singularity/3.6.2                                  
#> [139] cellranger/7.0.0             minimap2/2.4               singularity/3.7.4                                  
#> [140] cellsnp-lite/1.2.0           minimap2/2.17              singularity/3.8.2                    
#> [141] circos/0.69                  molgenis-compute/17.08.1   singularity/3.8.4                                  
#> [142] circos/0.69-9                mono/6.10.0.104            snpEff/4.3i                                        
#> [143] clover/2015-08-28            mono/6.12.0.90             snpEff/4.3t            
#> [144] cnvnator/0.3.3               MultiQC/1.10.1             SNPsplit/0.2.0         
#> [145] computecpp/1.0.5             multiSNV/2.3-10-g06102a5   soapdenovo-trans/1.03  
#> [146] cutadapt/2.9                 muscle/3.8.31              soapdenovo2/2.04                     
#> [147] cutadapt/3.4                 mutect/1.1.7               spaceranger/1.2.2                                  
#> [148] cutadapt/3.7                 nanonet-openBLAS/2.0.0     spaceranger/1.3.0                    
#> [149] deeptools/3.5.1              nanonet/2.0.0              spaceranger/2.0.0      
#> [150] dotnet/2.1.809               nanopolish/0.5.0           spades/3.11.0                        
#> [151] dotnet/3.1.412               nanopolish/0.6.2           spawn-fcgi/1.6.5                                   
#> [152] duck/6.9.3                   nanopolish/0.7.1           spp/1.14                             
#> [153] edirect/15.0                 nanopolish/0.8.1           sra-toolkit/2.8.2-1                                
#> [154] egaclient/2.2.2              nanopolish/0.8.5           sra-toolkit/2.9.6-1                  
#> [155] ensembl-vep/89.0             nanopore/1.0               sra-toolkit/2.10.9                                 
#> [156] ensembl-vep/89.8             ncbi-blast/2.6.0+          sra-toolkit/3.0.0                                  
#> [157] ensembl-vep/98.3             ncbi-blast/2.7.1+          sshpass/1.05                         
#> [158] fastqc/0.11.5                ncbi-blast/2.9.0           stack/1.5.1                                                      
#> [159] fastqc/0.11.8                ncbi-blast/2.9.0a          stack/1.7.1                          
#> [160] fastx-toolkit/0.0.13         ncbi-blast/2.10.1+         STAR/2.6.1c                                        
#> [161] ffmpeg/3.4                   ncbi-blast/2.11.0+         STAR/2.7.0a                          
#> [162] flash/1.2.11                 nedit/5.5                  STAR/2.7.1a                          
#> [163] freebayes/1.1.0              netcdf/4.4.1.1             STAR/2.7.3a                          
#> [164] freebayes/1.2.0              nix/fix                    STAR/2.7.9a                          
#> [165] gatk/4.2.4.1                 node/6.11.0                stornext/1.0                                       
#> [166] gatk/4.2.5.0                 node/8.11.2                subread/1.5.0-p2                     
#> [167] gcc/6.3.0                    node/9.9.0                 subread/1.5.1                                                    
#> [168] gcc/7.1.0                    novoalign/3.04.04          subread/1.5.2                                      
#> [169] gcc/7.3.0                    novocraft/3.08.02          subread/1.6.3                        
#> [170] gcc/8.1.0                    oases/0.2.09               tadbit/0.2                           
#> [171] gcc/8.2.0                    pandoc/1.17.0.3            texlive/3.14159265                                               
#> [172] gcc/8.3.0                    pandoc/2.0.0.1             texlive/20171130                                   
#> [173] gcc/9.1.0                    pandoc/2.3.1               tmhmm/2.0c             
#> [174] gcc/10.3.0                   parallel/20190722          tophat/2.1.1                         
#> [175] gcc/11.1.0                   pbs-drmaa/1.0.19           transdecoder/4.0.0                   
#> [176] gcc/11.2.0                   pbstop/1.0                 TRF/4.09.1                           
#> [177] gem/3.0                      pcre2/10.34-gcc-9.1.0      trimgalore/0.4.2                     
#> [178] gemini/0.20.1                pcre2/10.36-gcc-11.1.0     trimgalore/0.4.4                                   
#> [179] gencore/0.17.2               peakseq/1.31               trimgalore/0.6.7       
#> [180] genetorrent/3.8.7            pear/0.9.10                trimmomatic/0.36       
#> [181] genozip/12.0.11              pear/0.9.11                trinityrnaseq/2.2.0                                
#> [182] geos/3.7.0                   perl/5.22.2                trinityrnaseq/2.4.0                  
#> [183] germline/1.5.1               perl/5.26.3                trinityrnaseq/2.8.4                  
#> [184] gh-cli/2.0                   perl/5.30.1                trinotate/3.0.1                                                  
#> [185] glib/1.2.10                  php/5.6.21                 trinotate/3.0.2        
#> [186] gmap-gsnap/2017-05-08        picard-tools/2.26.11       turbovnc/3.0.1                       
#> [187] gmap-gsnap/2017-06-16        pindel/0.2.5b9             ucsc-tools/331         
#> [188] google-cloud-sdk/194.0.0     pixz/1.0.6                 ultraplex/1.2.5        
#> [189] google-cloud-sdk/348.0.0     platypus/0.8.1             umi_tools/1.0.1                      
#> [190] gradle/4.3                   plink/1.9                  usearch/7.0.1090                     
#> [191] graphmap/0.3.1               poretools/0.6.0            usearch/8.1.1861                                   
#> [192] graphmap/0.5.1               python/2.7.18              usearch/9.2.64                                     
#> [193] graphmap/0.5.2               python/3.5.1               vardict/1.5.1                        
#> [194] gridss/2.11.1                python/3.5.3               varscan/2.3.9                        
#> [195] gsl/2.4                      python/3.6.5-intel         vcflib/1.0.0-rc1                                   
#> [196] gsl/2.6                      python/3.7.0               vcftools/0.1.13                      
#> [197] guppy-cpu/2.3.7              python/3.7.13              velvet/1.2.10                                      
#> [198] guppy-cpu/3.1.5              python/3.8.3               vireoSNP/0.5.6                                     
#> [199] guppy-cpu/3.2.1              python/3.8.8               vsearch/2.4.3                        
#> [200] guppy-cpu/3.2.2              python/3.9.5               wehi-stornext/1.0                                  
#> [201] guppy-cpu/3.2.4              python/3.10.4              xenomapper/1.0.1                                   
#> [202] guppy-cpu/3.3.0              qiime2/2019.7              Xvfb/1.19.5                          
#> [203] guppy-cpu/3.6.1              qiime2/2020.8              xz/5.2.2                                           
#> [204] guppy-cpu/4.0.15             R/3.3.3                    zlib/1.2.8                                         
#> [205] haplopainter/1.043           R/3.4.4                    zlib/1.2.11                          
#> [206] hdf4/4.2.11                  R/3.5.2                                                         
#> [207] hdf5/1.8.16                  R/3.5.3                                                                                     
#> [208]                                                                                                                          
#> [209] ------------------- /stornext/System/data/modulefiles/nvidia -------------------                        
#> [210] clara-parabricks/4.0.1-1   guppy-gpu/6.2.1                topaz/0.2.3                        
#> [211] crYOLO/1.7.5cpu            libglvnd/1.1.1                 topaz/0.2.4                        
#> [212] crYOLO/1.7.5gpu            libtorch-gpu/1.12.1                                               
#> [213] crYOLO/1.8.0gpu            magma-gpu/2.7.0-11.3.1                                                          
#> [214] ctffind/4.1.5              ModelAngelo/0.0.1                                                               
#> [215] ctffind/4.1.13             motioncor2/1.2.1cu8.0                                                                         
#> [216] ctffind/4.1.14             motioncor2/1.2.1cu9.2                                                           
#> [217] cuda/8.0                   motioncor2/1.2.2cu8.0                                                                         
#> [218] cuda/9.2                   motioncor2/1.2.2cu9.2                                                                         
#> [219] cuda/10.0                  motioncor2/1.2.2cu10.0                                                          
#> [220] cuda/10.1                  motioncor2/1.2.6cu8.0                                                                         
#> [221] cuda/10.2                  motioncor2/1.2.6cu9.2                                                                         
#> [222] cuda/11.0                  motioncor2/1.2.6cu10.1                                                                        
#> [223] cuda/11.1                  motioncor2/1.3.1cu9.2                                                                         
#> [224] cuda/11.2                  motioncor2/1.3.1cu10.1                                                                        
#> [225] cuda/11.3.1                motioncor2/1.4.0cu9.2                                                                         
#> [226] cuda/11.4                  motioncor2/1.4.0cu10.0                                                                        
#> [227] cuda/11.5.1                motioncor2/1.4.0cu10.1                                                                        
#> [228] cuda/11.6.2                motioncor2/1.4.0cu10.2                                                                        
#> [229] cuda/11.7.1                motioncor2/1.4.0cu11.0                                                                        
#> [230] CUDA8/caffe2/0.8.1         motioncor2/1.4.4cu10.0                                                          
#> [231] CUDA8/tensorflow/1.1.0     motioncor2/1.4.4cu10.1                                                          
#> [232] CUDA8/tensorflow/1.3.0rc0  motioncor2/1.4.4cu10.2                                                                        
#> [233] cudnn/10.1-7.6.5           motioncor2/1.4.4cu11.1                                                                        
#> [234] cudnn/11.2-8.1.1.33        motioncor2/1.4.4cu11.2                                                                        
#> [235] cudnn/11.3.1-8.4.0         motioncor2/1.4.4cu11.3.1                                                                      
#> [236] cudnn/11.3.1-8.6.0.162     motioncor2/1.5.0cu10.1                                                          
#> [237] darknet/61c9d02cu10.0      motioncor2/1.5.0cu10.2                                                          
#> [238] deepEMhancer/0.15          motioncor2/1.5.0cu11.1                                                          
#> [239] dorado/0.1.1-gcc-9.1.0     motioncor2/1.5.0cu11.2                                                          
#> [240] EMAN/2.3                   motioncor2/1.5.0cu11.3                                                          
#> [241] EMAN2/2.9                  motioncor2/1.5.0cu11.4                                                          
#> [242] gautomatch/0.56cu8.0       motioncorr/2.1                                                                  
#> [243] gautomatch/0.56cu10.1      NAMD/2.12                                                                       
#> [244] Gctf/1.06                  NAMD/2.13nb                                                                     
#> [245] Gctf/1.18cu7.5             openmpi-CUDA8/3.0.0                                                             
#> [246] Gctf/1.18cu8.0             relion/3.1.3-cu11.2                                                             
#> [247] Gctf/1.18cu9.2             relion/3.1.3-gcc11.1.0-cu11.4                                                   
#> [248] Gctf/1.18cu10.1            relion/4.0-beta-1-cu10.1                                                                      
#> [249] guppy-gpu/2.3.1            relion/4.0-beta-1-cu11.2                                                                      
#> [250] guppy-gpu/2.3.7            scipion/2.0                                                                                   
#> [251] guppy-gpu/3.1.5            scipion/3.0                                                                     
#> [252] guppy-gpu/3.2.1            tensorflow-gpu/10.0-1.15.2                                                                    
#> [253] guppy-gpu/3.2.2            tensorflow-gpu/10.1-2.1.0                                                       
#> [254] guppy-gpu/3.2.4            tensorflow-gpu/11.2-2.11.0                                                                    
#> [255] guppy-gpu/3.3.0            tensorRT/10.0-6.0.1.5                                                                         
#> [256] guppy-gpu/3.6.1            tensorRT/10.1-6.0.1.5                                                           
#> [257] guppy-gpu/4.0.15           tensorRT/11.2-7.2.3.4                                                                         
#> [258]                                                                                                                          
#> [259] --------------- /stornext/System/data/modulefiles/structbio/its ----------------                        
#> [260] adxv/1.9.12        coot/0.8.8                phenix-rosetta/1.11.1-2575                      
#> [261] alphafold/2.0.0    coot/0.8.9                phenix/1.15-3459                                
#> [262] alphafold/2.0.1    coot/0.8.9.1              phenix/1.17-3644                                                            
#> [263] alphafold/2.0.2    cryoEF/1.1.0              phenix/1.18.2-3874                                            
#> [264] alphafold/2.1.1.0  dalilite/5.1              phenix/1.20.1-4487                              
#> [265] alphafold/2.1.2.0  dials/3.10.0              pyem/0.5                                                      
#> [266] alphafold/2.2.0.0  dials/3.12.0              pymol/1.7.4.0                                                 
#> [267] autoPROC/20211020  eiger2cbf/160415          pymol/2.5.0                                     
#> [268] autoPROC/20220608  EMBuilder/1.0             relion-cpu/3.1b-icc                             
#> [269] autoSHARP/2.8      EPU_group_AFIS/0.3        relion-cpu/3.1beta-icc            
#> [270] bsoft/2.1.0        glew/1.9.0                schrodinger/2022-1                              
#> [271] buster/20211020    gromacs/4.5.5_cryo_fit    schrodinger/2022-2                                            
#> [272] buster/20220608    gromacs/2021.2            schrodinger/2022-3                              
#> [273] ccp4/7.1           H5toXDS/1.1.0             sidesplitter/1.2                  
#> [274] ccp4/8.0           IMOD/4.11.5               SPHIRE/1.1                                      
#> [275] ccpem/1.5.0        mrcz/0.1.3                SPHIRE/1.3                                      
#> [276] chimera/1.11       NAMD-mpich/2.12           summovie/1.0.2                    
#> [277] chimerax/0.8       NAMD-mpich/2.14           tvips/0.0.3                                                   
#> [278] chimerax/0.9       NAMD-multicore-gpu/2.14   unblur/1.0.2                                                  
#> [279] chimerax/0.93      NAMD-multicore/2.12       xds/20220110                                    
#> [280] chimerax/1.1       NAMD-multicore/2.14       xdsgui/2017-01-25                               
#> [281] chimerax/1.3       NAMD-openmpi/2.12         xdsgui/2020-12-02                                             
#> [282] cistem/1.0.0-beta  phenix-nightly/1.15-3459                                                  
#> [283]                                                                                                                          
#> [284] Key:                                                                                                                     
#> [285] modulepath  default-version